+++ src/typechecker/tests.rs
@@ -10,3278 +10,9 @@ fn parse_program(src: &str) -> ast::Program {
     crate::parser::parse(tokens, file_id).expect("parse")
 }
 
-#[test]
-fn resolve_primitive_types() {
-    let tc = TypeChecker::new();
-    assert_eq!(tc.resolve_type(&AstType::I32), Type::I32);
-    assert_eq!(tc.resolve_type(&AstType::F64), Type::F64);
-    assert_eq!(tc.resolve_type(&AstType::Bool), Type::Bool);
-    assert_eq!(tc.resolve_type(&AstType::Void), Type::Void);
-    assert_eq!(tc.resolve_type(&AstType::Str), Type::Str);
-}
+mod resolver_metadata;
 
-#[test]
-fn resolve_pointer_types() {
-    let tc = TypeChecker::new();
-    assert_eq!(
-        tc.resolve_type(&AstType::Ptr(Box::new(AstType::I32))),
-        Type::Ptr(Box::new(Type::I32))
-    );
-}
-
-#[test]
-fn method_signature_key_helpers_share_receiver_parsing() {
-    assert_eq!(method_signature_key("Point", "get"), "Point.get");
-    assert_eq!(
-        method_signature_key_parts("Point.get"),
-        Some(("Point", "get"))
-    );
-    assert_eq!(method_signature_receiver_name("Point.get"), Some("Point"));
-    assert_eq!(
-        method_signature_method_name_for_receiver("Point.get", "Point"),
-        Some("get")
-    );
-    assert_eq!(
-        method_signature_method_name_for_receiver("Other.get", "Point"),
-        None
-    );
-    assert!(is_method_signature_key("Point.get"));
-    assert_eq!(method_signature_key_parts("plain"), None);
-    assert_eq!(method_signature_receiver_name("plain"), None);
-    assert!(!is_method_signature_key("plain"));
-}
-
-#[test]
-fn resolver_symbol_lookup_helpers_share_definition_span_fallbacks() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Point.get = (self: Point) i32 { self.x }
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let Declaration::Method { span, .. } = &program.declarations[1] else {
-        panic!("expected method declaration");
-    };
-    let span = *span;
-
-    assert_eq!(
-        TypeChecker::resolver_symbol_name_for(&symbols, Namespace::Value, "Point.missing", span),
-        "Point.get"
-    );
-    assert_eq!(
-        TypeChecker::resolver_method_signature_name_for(
-            &symbols,
-            "Missing.missing",
-            "Missing",
-            span
-        ),
-        "Point.get"
-    );
-    assert_eq!(
-        TypeChecker::resolver_method_signature_symbol_by_span(&symbols, span)
-            .map(|symbol| symbol.name.as_str()),
-        Some("Point.get")
-    );
-}
-
-#[test]
-fn resolver_count_display_formats_known_and_missing_counts() {
-    assert_eq!(resolver_count_display(Some(2)), "2");
-    assert_eq!(resolver_count_display(None), "unknown");
-}
-
-#[test]
-fn count_validation_formats_message() {
-    let validation = CountValidation {
-        label: "parameter count",
-        code: "COUNT",
-    };
-
-    assert_eq!(validation.code, "COUNT");
-    assert_eq!(
-        validation.message("value", "add", Some(1), 2),
-        "resolver value symbol 'add' has parameter count 1, expected 2"
-    );
-    assert_eq!(
-        validation.message("variant", "Some", None, 1),
-        "resolver variant symbol 'Some' has parameter count unknown, expected 1"
-    );
-}
-
-#[test]
-fn count_validation_uses_value_parameter_resolver_code() {
-    let validation = CountValidation::value_parameter_resolver_code();
-
-    assert_eq!(validation.label, "parameter count");
-    assert_eq!(validation.code, "E0211");
-}
-
-#[test]
-fn count_validation_uses_field_resolver_code() {
-    let validation = CountValidation::field_resolver_code();
-
-    assert_eq!(validation.label, "field count");
-    assert_eq!(validation.code, "E0214");
-}
-
-#[test]
-fn count_validation_uses_variant_payload_resolver_code() {
-    let validation = CountValidation::variant_payload_resolver_code();
-
-    assert_eq!(validation.label, "payload count");
-    assert_eq!(validation.code, "E0215");
-}
-
-#[test]
-fn type_parameter_validation_formats_messages() {
-    let validation = TypeParameterValidation {
-        count_code: "COUNT",
-        name_code: "NAMES",
-        bound_code: "BOUNDS",
-        bound_ref_code: "BOUND_REFS",
-    };
-
-    assert_eq!(validation.name_code, "NAMES");
-    assert_eq!(
-        validation.name_message("value", "identity", "(U)", "(T)"),
-        "resolver value symbol 'identity' has type parameter names '(U)', expected '(T)'"
-    );
-    assert_eq!(
-        validation.bound_message("type", "Box", "(T: Other)", "(T: Json)"),
-        "resolver type symbol 'Box' has type parameter bounds '(T: Other)', expected '(T: Json)'"
-    );
-    assert_eq!(
-            validation.bound_ref_message("behavior", "Serializable", "(T: Json<i32>)", "(T: Json<T>)"),
-            "resolver behavior symbol 'Serializable' has type parameter bound refs '(T: Json<i32>)', expected '(T: Json<T>)'"
-        );
-}
-
-#[test]
-fn type_parameter_validation_uses_type_like_resolver_codes() {
-    let validation = TypeParameterValidation::type_like_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0213");
-    assert_eq!(validation.name_code, "E0346");
-    assert_eq!(validation.bound_code, "E0222");
-    assert_eq!(validation.bound_ref_code, "E0350");
-}
-
-#[test]
-fn type_parameter_validation_uses_value_resolver_codes() {
-    let validation = TypeParameterValidation::value_resolver_codes();
-
-    assert_eq!(validation.count_code, "E0220");
-    assert_eq!(validation.name_code, "E0347");
-    assert_eq!(validation.bound_code, "E0221");
-    assert_eq!(validation.bound_ref_code, "E0351");
-}
-
-#[test]
-fn type_parameter_validation_builds_count_validation() {
-    let validation = TypeParameterValidation {
-        count_code: "COUNT",
-        name_code: "NAMES",
-        bound_code: "BOUNDS",
-        bound_ref_code: "BOUND_REFS",
-    }
-    .count_validation();
-
-    assert_eq!(validation.label, "type parameter count");
-    assert_eq!(validation.code, "COUNT");
-}
-
-#[test]
-fn value_parameter_validation_formats_messages() {
-    let validation = ValueParameterValidation {
-        name_code: "NAMES",
-        display_type_code: "TYPES",
-        typed_type_code: "TYPED_TYPES",
-    };
-
-    assert_eq!(validation.name_code, "NAMES");
-    assert_eq!(
-        validation.name_message("add", "(a, other)", "(a, b)"),
-        "resolver value symbol 'add' has parameter names '(a, other)', expected '(a, b)'"
-    );
-    assert_eq!(
-        validation.display_type_message("add", "(i32, i32)", "(i32, f64)"),
-        "resolver value symbol 'add' has parameter types '(i32, i32)', expected '(i32, f64)'"
-    );
-    assert_eq!(
-        validation.typed_type_message("apply", "(i32)", "((i32) i32)"),
-        "resolver value symbol 'apply' has typed parameter types '(i32)', expected '((i32) i32)'"
-    );
-}
-
-#[test]
-fn value_parameter_validation_uses_resolver_codes() {
-    let validation = ValueParameterValidation::resolver_codes();
-
-    assert_eq!(validation.name_code, "E0223");
-    assert_eq!(validation.display_type_code, "E0216");
-    assert_eq!(validation.typed_type_code, "E0356");
-}
-
-#[test]
-fn return_validation_formats_messages() {
-    let validation = ReturnValidation {
-        display_code: "RETURN",
-        typed_code: "TYPED_RETURN",
-    };
-
-    assert_eq!(validation.display_code, "RETURN");
-    assert_eq!(
-        validation.display_message("main", "bool", "i32"),
-        "resolver value symbol 'main' has return type 'bool', expected 'i32'"
-    );
-    assert_eq!(
-        validation.typed_message("apply", "i32", "(i32) i32"),
-        "resolver value symbol 'apply' has typed return type 'i32', expected '(i32) i32'"
-    );
-}
-
-#[test]
-fn return_validation_uses_resolver_codes() {
-    let validation = ReturnValidation::resolver_codes();
-
-    assert_eq!(validation.display_code, "E0212");
-    assert_eq!(validation.typed_code, "E0357");
-}
-
-#[test]
-fn behavior_method_validation_formats_messages() {
-    let validation = BehaviorMethodValidation {
-        display_code: "METHODS",
-        typed_code: "TYPED_METHODS",
-    };
-
-    assert_eq!(validation.display_code, "METHODS");
-    assert_eq!(
-            validation.display_message("Serializable", "(encode(Self) bool)", "(encode(Self) str)"),
-            "resolver behavior symbol 'Serializable' has methods '(encode(Self) bool)', expected '(encode(Self) str)'"
-        );
-    assert_eq!(
-            validation.typed_message(
-                "Mapper",
-                "(map(__arg0: Self, __arg1: i32) i32)",
-                "(map(__arg0: Self, __arg1: (i32) i32) i32)"
-            ),
-            "resolver behavior symbol 'Mapper' has typed methods '(map(__arg0: Self, __arg1: i32) i32)', expected '(map(__arg0: Self, __arg1: (i32) i32) i32)'"
-        );
-}
-
-#[test]
-fn behavior_method_validation_uses_resolver_codes() {
-    let validation = BehaviorMethodValidation::resolver_codes();
-
-    assert_eq!(validation.display_code, "E0219");
-    assert_eq!(validation.typed_code, "E0355");
-}
-
-#[test]
-fn field_validation_formats_messages() {
-    let validation = FieldValidation {
-        display_code: "FIELDS",
-        typed_code: "TYPED_FIELDS",
-    };
-
-    assert_eq!(validation.display_code, "FIELDS");
-    assert_eq!(
-        validation.display_message("type", "Point", "(x: i32)", "(x: f64)"),
-        "resolver type symbol 'Point' has fields '(x: i32)', expected '(x: f64)'"
-    );
-    assert_eq!(
-            validation.typed_message("type", "Pipeline", "(callback: i32)", "(callback: (i32) i32)"),
-            "resolver type symbol 'Pipeline' has typed fields '(callback: i32)', expected '(callback: (i32) i32)'"
-        );
-}
-
-#[test]
-fn field_validation_uses_resolver_codes() {
-    let validation = FieldValidation::resolver_codes();
-
-    assert_eq!(validation.display_code, "E0217");
-    assert_eq!(validation.typed_code, "E0358");
-}
-
-#[test]
-fn variant_payload_validation_formats_messages() {
-    let validation = VariantPayloadValidation {
-        display_code: "PAYLOAD",
-        typed_code: "TYPED_PAYLOAD",
-    };
-
-    assert_eq!(validation.display_code, "PAYLOAD");
-    assert_eq!(
-        validation.display_message("Some", "bool", "i32"),
-        "resolver variant symbol 'Some' has payload type 'bool', expected 'i32'"
-    );
-    assert_eq!(
-        validation.typed_message("Wrap", "i32", "(i32) i32"),
-        "resolver variant symbol 'Wrap' has typed payload type 'i32', expected '(i32) i32'"
-    );
-}
-
-#[test]
-fn variant_payload_validation_uses_resolver_codes() {
-    let validation = VariantPayloadValidation::resolver_codes();
-
-    assert_eq!(validation.display_code, "E0218");
-    assert_eq!(validation.typed_code, "E0359");
-}
-
-#[test]
-fn variant_owner_validation_formats_message() {
-    let validation = VariantOwnerValidation { code: "OWNER" };
-
-    assert_eq!(validation.code, "OWNER");
-    assert_eq!(
-        validation.message("Some", "Result", "Option"),
-        "resolver variant symbol 'Some' has owner 'Result', expected 'Option'"
-    );
-}
-
-#[test]
-fn variant_owner_validation_uses_resolver_code() {
-    let validation = VariantOwnerValidation::resolver_code();
-
-    assert_eq!(validation.code, "E0242");
-}
-
-#[test]
-fn variant_name_validation_formats_message() {
-    let validation = VariantNameValidation { code: "VARIANTS" };
-
-    assert_eq!(validation.code, "VARIANTS");
-    assert_eq!(
-        validation.message("Option", "(Some)", "(Some, None)"),
-        "resolver type symbol 'Option' has variants '(Some)', expected '(Some, None)'"
-    );
-}
-
-#[test]
-fn variant_name_validation_uses_resolver_code() {
-    let validation = VariantNameValidation::resolver_code();
-
-    assert_eq!(validation.code, "E0241");
-}
-
-#[test]
-fn resolver_metadata_display_formats_known_and_missing_values() {
-    assert_eq!(resolver_metadata_display(Some("Point")), "Point");
-    assert_eq!(resolver_metadata_display(None), "unknown");
-    assert_eq!(
-        resolver_ast_type_metadata_display(Some(&AstType::I32)),
-        "i32"
-    );
-    assert_eq!(resolver_ast_type_metadata_display(None), "unknown");
-    assert_eq!(
-        optional_ast_type_display(Some(&AstType::Bool), "none"),
-        "bool"
-    );
-    assert_eq!(optional_ast_type_display(None, "none"), "none");
-}
-
-#[test]
-fn resolver_string_list_display_formats_known_and_missing_lists() {
-    let names = vec!["T".to_string(), "U".to_string()];
-    assert_eq!(join_resolver_strings(&names), "T, U");
-    assert_eq!(
-        join_resolver_display_values(&[AstType::I32, AstType::Bool], AstType::display_name),
-        "i32, bool"
-    );
-    assert_eq!(format_resolver_string_list(Some(&names)), "(T, U)");
-    assert_eq!(format_resolver_string_list(None), "unknown");
-}
-
-#[test]
-fn resolver_display_list_formats_mapped_known_and_missing_items() {
-    let types = vec![AstType::I32, AstType::Bool];
-    assert_eq!(format_ast_type_list(Some(&types)), "(i32, bool)");
-    assert_eq!(format_ast_type_list(None), "unknown");
-
-    let bounds = vec![("T".to_string(), "Display".to_string())];
-    assert_eq!(format_type_parameter_bounds(Some(&bounds)), "(T: Display)");
-    assert_eq!(format_type_parameter_bounds(None), "unknown");
-
-    let bound_refs = vec![TypeParameterBoundRefMetadata {
-        type_parameter: "T".to_string(),
-        behavior: "Display".to_string(),
-        type_args: vec![AstType::I32],
-    }];
-    assert_eq!(
-        format_type_parameter_bound_refs(Some(&bound_refs)),
-        "(T: Display<i32>)"
-    );
-    assert_eq!(format_type_parameter_bound_refs(None), "unknown");
-}
-
-#[test]
-fn resolver_nonempty_joined_list_formats_present_empty_and_missing_items() {
-    let names = vec!["Json".to_string(), "Debug".to_string()];
-    assert_eq!(format_behavior_ref_names(Some(&names)), "Json, Debug");
-    assert_eq!(format_behavior_ref_names(Some(&[])), "none");
-    assert_eq!(format_behavior_ref_names(None), "none");
-
-    let refs = vec![BehaviorRefMetadata {
-        name: "Json".to_string(),
-        type_args: vec![AstType::I32],
-    }];
-    assert_eq!(format_behavior_refs(Some(&refs)), "Json<i32>");
-    assert_eq!(format_behavior_refs(Some(&[])), "none");
-    assert_eq!(format_behavior_refs(None), "none");
-}
-
-#[test]
-fn resolver_behavior_ref_helpers_share_pop_and_peek_selection() {
-    let refs = VecDeque::from(vec![
-        BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![AstType::I32],
-        },
-        BehaviorRefMetadata {
-            name: "Debug".to_string(),
-            type_args: vec![],
-        },
-    ]);
-    let mut refs_by_type = HashMap::from([("Point".to_string(), refs.clone())]);
-
-    assert_eq!(
-        TypeChecker::peek_resolver_behavior_ref(true, &refs_by_type, "Point", "Debug")
-            .map(|reference| reference.name.as_str()),
-        Some("Debug")
-    );
-    assert_eq!(
-        TypeChecker::pop_resolver_behavior_ref(true, &mut refs_by_type, "Point", "Debug")
-            .map(|reference| reference.name),
-        Some("Debug".to_string())
-    );
-
-    let mut refs_by_type = HashMap::from([("Point".to_string(), refs)]);
-    assert_eq!(
-        TypeChecker::peek_resolver_behavior_ref(true, &refs_by_type, "Point", "Missing")
-            .map(|reference| reference.name.as_str()),
-        Some("Json")
-    );
-    assert_eq!(
-        TypeChecker::pop_resolver_behavior_ref(true, &mut refs_by_type, "Point", "Missing")
-            .map(|reference| reference.name),
-        Some("Json".to_string())
-    );
-    assert!(
-        TypeChecker::peek_resolver_behavior_ref(false, &refs_by_type, "Point", "Debug").is_none()
-    );
-    assert!(
-        TypeChecker::pop_resolver_behavior_ref(false, &mut refs_by_type, "Point", "Debug")
-            .is_none()
-    );
-}
-
-#[test]
-fn resolver_behavior_ref_for_selects_impl_and_required_queues_by_role() {
-    let mut tc = TypeChecker::new();
-    tc.resolver_backed_collection = true;
-    tc.resolver_behavior_impl_refs.insert(
-        "Point".to_string(),
-        VecDeque::from([BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![AstType::I32],
-        }]),
-    );
-    tc.resolver_behavior_required_refs.insert(
-        "Point".to_string(),
-        VecDeque::from([BehaviorRefMetadata {
-            name: "Debug".to_string(),
-            type_args: vec![],
-        }]),
-    );
-
-    assert_eq!(
-        tc.resolver_behavior_ref_for(BehaviorRefRole::Impl, "Point", "Json")
-            .map(|reference| reference.name),
-        Some("Json".to_string())
-    );
-    assert_eq!(
-        tc.resolver_behavior_ref_for(BehaviorRefRole::Required, "Point", "Debug")
-            .map(|reference| reference.name),
-        Some("Debug".to_string())
-    );
-    assert!(tc
-        .resolver_behavior_ref_for(BehaviorRefRole::Parent, "Point", "Json")
-        .is_none());
-
-    tc.resolver_behavior_impl_refs.insert(
-        "Point".to_string(),
-        VecDeque::from([BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![],
-        }]),
-    );
-    tc.resolver_backed_collection = false;
-    assert!(tc
-        .resolver_behavior_ref_for(BehaviorRefRole::Impl, "Point", "Json")
-        .is_none());
-}
-
-#[test]
-fn resolver_params_from_metadata_preserves_ast_param_shape() {
-    let existing_span = Span::new(1, 10, 15);
-    let default_span = Span::new(1, 20, 25);
-    let existing_params = vec![Param {
-        name: "stale".to_string(),
-        ty: AstType::Bool,
-        mutable: true,
-        span: existing_span,
-    }];
-    let parameter_names = vec!["value".to_string(), "mapper".to_string()];
-    let parameter_types = vec![
-        AstType::I32,
-        AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        },
-    ];
-
-    let params = TypeChecker::resolver_params_from_metadata(
-        &existing_params,
-        &parameter_names,
-        &parameter_types,
-        default_span,
-    );
-
-    assert_eq!(params[0].name, "value");
-    assert_eq!(params[0].ty, AstType::I32);
-    assert!(params[0].mutable);
-    assert_eq!(params[0].span, existing_span);
-    assert_eq!(params[1].name, "mapper");
-    assert_eq!(params[1].ty, parameter_types[1]);
-    assert!(!params[1].mutable);
-    assert_eq!(params[1].span, default_span);
-}
-
-#[test]
-fn resolver_optional_return_type_maps_void_to_missing_annotation() {
-    assert_eq!(
-        TypeChecker::resolver_optional_return_type(&AstType::Void),
-        None
-    );
-    assert_eq!(
-        TypeChecker::resolver_optional_return_type(&AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        }),
-        Some(AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        })
-    );
-}
-
-#[test]
-fn resolver_enum_variants_from_metadata_uses_owner_scoped_payloads() {
-    let program = parse_program(
-        r#"
-First: Wrap(i32), None
-Second: Wrap(str)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let variants = vec!["Wrap".to_string(), "None".to_string()];
-
-    assert_eq!(
-        TypeChecker::resolver_enum_variants_from_metadata(&symbols, "First", &variants),
-        vec![
-            ("Wrap".to_string(), Some(AstType::I32)),
-            ("None".to_string(), None),
-        ]
-    );
-    assert_eq!(
-        TypeChecker::resolver_enum_variants_from_metadata(
-            &symbols,
-            "Second",
-            &["Wrap".to_string()]
-        ),
-        vec![("Wrap".to_string(), Some(AstType::Str))]
-    );
-}
-
-#[test]
-fn resolver_struct_fields_from_metadata_restores_field_names_and_defaults() {
-    let fields = vec![
-        ("x".to_string(), AstType::I32),
-        (
-            "callback".to_string(),
-            AstType::Function {
-                params: vec![AstType::I32],
-                ret: Box::new(AstType::Str),
-            },
-        ),
-    ];
-    let ast_fields = vec![
-        StructField {
-            name: "stale_x".to_string(),
-            ty: AstType::Bool,
-            default: Some(Expression::BoolLiteral {
-                value: true,
-                span: Span::dummy(),
-            }),
-            mutable: false,
-            span: Span::dummy(),
-        },
-        StructField {
-            name: "stale_callback".to_string(),
-            ty: AstType::Bool,
-            default: None,
-            mutable: false,
-            span: Span::dummy(),
-        },
-    ];
-
-    let (restored_fields, defaults) =
-        TypeChecker::resolver_struct_fields_from_metadata(&fields, &ast_fields);
-
-    assert_eq!(restored_fields, fields);
-    assert!(defaults.contains_key("x"));
-    assert!(!defaults.contains_key("stale_x"));
-    assert!(!defaults.contains_key("callback"));
-}
-
-#[test]
-fn resolver_behavior_methods_from_metadata_preserves_defaults_by_resolver_order() {
-    let first_span = Span::new(1, 10, 15);
-    let second_span = Span::new(1, 20, 25);
-    let default_span = Span::new(1, 30, 35);
-    let existing_methods = vec![
-        ast::BehaviorMethod {
-            name: "first".to_string(),
-            params: vec![Param {
-                name: "stale".to_string(),
-                ty: AstType::Bool,
-                mutable: false,
-                span: first_span,
-            }],
-            return_type: Some(AstType::Bool),
-            default_body: Some(Expression::IntLiteral {
-                value: 1,
-                span: first_span,
-            }),
-            span: first_span,
-        },
-        ast::BehaviorMethod {
-            name: "second".to_string(),
-            params: vec![],
-            return_type: None,
-            default_body: Some(Expression::IntLiteral {
-                value: 2,
-                span: second_span,
-            }),
-            span: second_span,
-        },
-    ];
-    let method_types = vec![
-        BehaviorMethodTypeMetadata {
-            name: "second".to_string(),
-            parameter_names: vec!["value".to_string()],
-            parameter_types: vec![AstType::I32],
-            return_type: AstType::Str,
-        },
-        BehaviorMethodTypeMetadata {
-            name: "first".to_string(),
-            parameter_names: vec![],
-            parameter_types: vec![],
-            return_type: AstType::Void,
-        },
-    ];
-
-    let methods = TypeChecker::resolver_behavior_methods_from_metadata(
-        existing_methods,
-        &method_types,
-        default_span,
-    );
-
-    assert_eq!(methods[0].name, "second");
-    assert_eq!(methods[0].params[0].name, "value");
-    assert_eq!(methods[0].params[0].ty, AstType::I32);
-    assert_eq!(methods[0].return_type, Some(AstType::Str));
-    assert_eq!(methods[0].span, second_span);
-    assert!(matches!(
-        methods[0].default_body,
-        Some(Expression::IntLiteral { value: 2, .. })
-    ));
-    assert_eq!(methods[1].name, "first");
-    assert_eq!(methods[1].return_type, None);
-    assert_eq!(methods[1].span, first_span);
-    assert!(matches!(
-        methods[1].default_body,
-        Some(Expression::IntLiteral { value: 1, .. })
-    ));
-}
-
-#[test]
-fn behavior_parent_refs_from_metadata_restores_keys_and_type_args() {
-    let tc = TypeChecker::new();
-    let metadata = vec![
-        BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![AstType::Named("T".to_string())],
-        },
-        BehaviorRefMetadata {
-            name: "Debug".to_string(),
-            type_args: vec![],
-        },
-    ];
-
-    let refs = tc.behavior_parent_refs_from_metadata(&metadata);
-
-    assert_eq!(refs[0].behavior, "Json");
-    assert_eq!(refs[0].type_args, vec![AstType::Named("T".to_string())]);
-    assert_eq!(refs[0].key, "Json_T");
-    assert_eq!(refs[1].behavior, "Debug");
-    assert!(refs[1].type_args.is_empty());
-    assert_eq!(refs[1].key, "Debug");
-}
-
-#[test]
-fn behavior_impl_refs_from_metadata_restores_type_and_behavior_keys() {
-    let tc = TypeChecker::new();
-    let metadata = vec![
-        BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![AstType::Str],
-        },
-        BehaviorRefMetadata {
-            name: "Debug".to_string(),
-            type_args: vec![],
-        },
-    ];
-
-    assert_eq!(
-        tc.behavior_impl_refs_from_metadata("Point", &metadata),
-        vec![
-            ("Point".to_string(), "Json_str".to_string()),
-            ("Point".to_string(), "Debug".to_string()),
-        ]
-    );
-}
-
-#[test]
-fn resolver_behavior_ref_queue_selection_prefers_exact_then_front() {
-    let refs = VecDeque::from(vec![
-        BehaviorRefMetadata {
-            name: "Json".to_string(),
-            type_args: vec![AstType::I32],
-        },
-        BehaviorRefMetadata {
-            name: "Debug".to_string(),
-            type_args: vec![],
-        },
-    ]);
-
-    assert_eq!(
-        TypeChecker::resolver_behavior_ref_queue_index(&refs, "Debug"),
-        Some(1)
-    );
-    assert_eq!(
-        TypeChecker::resolver_behavior_ref_queue_index(&refs, "Missing"),
-        Some(0)
-    );
-    assert_eq!(
-        TypeChecker::resolver_behavior_ref_queue_index(&VecDeque::new(), "Missing"),
-        None
-    );
-}
-
-#[test]
-fn named_queue_selection_prefers_exact_then_front() {
-    let items = VecDeque::from(["Json".to_string(), "Debug".to_string()]);
-
-    assert_eq!(
-        TypeChecker::named_queue_index(&items, "Debug", String::as_str),
-        Some(1)
-    );
-    assert_eq!(
-        TypeChecker::named_queue_index(&items, "Missing", String::as_str),
-        Some(0)
-    );
-    assert_eq!(
-        TypeChecker::named_queue_index(&VecDeque::<String>::new(), "Missing", String::as_str),
-        None
-    );
-}
-
-#[test]
-fn named_queue_selection_can_preserve_front_for_future_match() {
-    let items = VecDeque::from(["Json".to_string(), "Debug".to_string()]);
-
-    assert_eq!(
-        TypeChecker::named_queue_index_preserving_future_front(
-            &items,
-            "Debug",
-            Vec::<&str>::new(),
-            String::as_str,
-        ),
-        Some(1)
-    );
-    assert_eq!(
-        TypeChecker::named_queue_index_preserving_future_front(
-            &items,
-            "Missing",
-            ["Json"],
-            String::as_str,
-        ),
-        None
-    );
-    assert_eq!(
-        TypeChecker::named_queue_index_preserving_future_front(
-            &items,
-            "Missing",
-            ["Other"],
-            String::as_str,
-        ),
-        Some(0)
-    );
-}
-
-#[test]
-fn impl_effective_method_name_prefers_resolver_then_ast_then_collected_signature() {
-    let mut tc = TypeChecker::new();
-    tc.resolver_backed_collection = true;
-    tc.methods.insert(
-        "Point.describe".to_string(),
-        FuncInfo {
-            name: "Point.describe".to_string(),
-            params: Vec::new(),
-            return_type: AstType::Void,
-            type_params: Vec::new(),
-            type_param_bounds: HashMap::new(),
-        },
-    );
-    let mut unmatched = VecDeque::from([
-        "encode".to_string(),
-        "debug".to_string(),
-        "describe".to_string(),
-    ]);
-
-    assert_eq!(
-        tc.impl_effective_method_name(
-            &mut unmatched,
-            "stale",
-            Some("Point.encode".to_string()),
-            "Point",
-        ),
-        "encode"
-    );
-    assert_eq!(
-        tc.impl_effective_method_name(&mut unmatched, "debug", None, "Point"),
-        "debug"
-    );
-    assert_eq!(
-        tc.impl_effective_method_name(&mut unmatched, "missing", None, "Point"),
-        "describe"
-    );
-    assert!(unmatched.is_empty());
-}
-
-#[test]
-fn resolver_backed_impl_method_key_requires_resolver_collection() {
-    let mut program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let (span, ast_key) = if let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &mut program.declarations[2]
-    {
-        *type_name = "Missing".to_string();
-        if let Declaration::Function { name, span, .. } = &mut methods[0] {
-            *name = "missing".to_string();
-            (*span, TypeChecker::method_key(type_name, name))
-        } else {
-            panic!("expected impl method");
-        }
-    } else {
-        panic!("expected impl block");
-    };
-    let mut tc = TypeChecker::new();
-
-    assert_eq!(
-        tc.resolver_backed_impl_method_key(Some(&symbols), &ast_key, "Missing", span),
-        None
-    );
-    tc.resolver_backed_collection = true;
-    assert_eq!(
-        tc.resolver_backed_impl_method_key(Some(&symbols), &ast_key, "Missing", span),
-        Some("Point.encode".to_string())
-    );
-}
-
-#[test]
-fn effective_behavior_impl_methods_carry_named_declaration_and_method_name() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let Declaration::ImplBlock { methods, .. } = &program.declarations[2] else {
-        panic!("expected impl block");
-    };
-    let mut tc = TypeChecker::new();
-    tc.resolver_backed_collection = true;
-    let mut unmatched = VecDeque::from(["encode".to_string()]);
-
-    let effective =
-        tc.effective_behavior_impl_methods(Some(&symbols), "Point", methods, &mut unmatched);
-
-    assert_eq!(effective.len(), 1);
-    assert_eq!(effective[0].method_name, "encode");
-    assert!(matches!(
-        effective[0].declaration,
-        Declaration::Function { name, .. } if name == "encode"
-    ));
-    assert!(unmatched.is_empty());
-}
-
-#[test]
-fn template_dependency_entries_use_named_fields() {
-    let entry = TemplateDependencyEntry::<StructInfo> {
-        name: "Point".to_string(),
-        previous: None,
-    };
-
-    assert_eq!(entry.name, "Point");
-    assert!(entry.previous.is_none());
-}
-
-#[test]
-fn resolver_backed_behavior_impl_method_signature_name_prefers_resolver_key() {
-    let tc = TypeChecker::new();
-    let mut required = VecDeque::from([
-        ast::BehaviorMethod {
-            name: "encode".to_string(),
-            params: Vec::new(),
-            return_type: Some(AstType::Str),
-            default_body: None,
-            span: Span::dummy(),
-        },
-        ast::BehaviorMethod {
-            name: "debug".to_string(),
-            params: Vec::new(),
-            return_type: Some(AstType::Str),
-            default_body: None,
-            span: Span::dummy(),
-        },
-    ]);
-
-    assert_eq!(
-        tc.resolver_backed_behavior_impl_method_signature_name(
-            &mut required,
-            "stale",
-            Some("Point.encode"),
-            "Point",
-        ),
-        Some("encode".to_string())
-    );
-    assert_eq!(
-        tc.resolver_backed_behavior_impl_method_signature_name(
-            &mut required,
-            "debug",
-            None,
-            "Point",
-        ),
-        Some("debug".to_string())
-    );
-    assert!(required.is_empty());
-}
-
-#[test]
-fn resolver_backed_method_signature_requires_resolver_collection() {
-    let mut tc = TypeChecker::new();
-    tc.methods.insert(
-        "Point.encode".to_string(),
-        FuncInfo {
-            name: "Point.encode".to_string(),
-            params: Vec::new(),
-            return_type: AstType::Str,
-            type_params: Vec::new(),
-            type_param_bounds: HashMap::new(),
-        },
-    );
-
-    assert!(tc
-        .resolver_backed_method_signature("Point", "encode")
-        .is_none());
-    tc.resolver_backed_collection = true;
-    assert_eq!(
-        tc.resolver_backed_method_signature("Point", "encode")
-            .map(|info| info.return_type.clone()),
-        Some(AstType::Str)
-    );
-}
-
-#[test]
-fn behavior_default_synthesis_skip_requires_resolver_collection_and_missing_impl_ref() {
-    let mut tc = TypeChecker::new();
-    tc.resolver_missing_behavior_impl_refs
-        .insert("Point".to_string());
-
-    assert!(!tc.should_skip_behavior_default_synthesis("Point"));
-    tc.resolver_backed_collection = true;
-    assert!(tc.should_skip_behavior_default_synthesis("Point"));
-    assert!(!tc.should_skip_behavior_default_synthesis("Other"));
-}
-
-#[test]
-fn resolver_backed_behavior_collection_defers_generic_metadata_to_resolver() {
-    let program = parse_program(
-        r#"
-Json<T: Json<T>>: behavior {
-    encode: (Self) T {
-        1
-    }
-}
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    tc.with_resolver_backed_collection(|checker| {
-        checker.collect_declarations(&program.declarations);
-    });
-
-    let behavior = tc.behaviors.get("Json").expect("behavior stub");
-    assert!(
-            behavior.type_params.is_empty(),
-            "resolver-backed behavior collection should not keep AST generic names before resolver metadata"
-        );
-    assert!(
-            behavior.type_param_bounds.is_empty(),
-            "resolver-backed behavior collection should not keep AST generic bounds before resolver metadata"
-        );
-    assert!(
-            behavior.methods[0].default_body.is_some(),
-            "resolver-backed behavior collection should still keep default bodies for later resolver metadata restoration"
-        );
-}
-
-#[test]
-fn method_key_formats_type_qualified_method_name() {
-    assert_eq!(TypeChecker::method_key("Point", "encode"), "Point.encode");
-}
-
-#[test]
-fn resolver_behavior_ref_owner_prefers_exact_then_unique_fallbacks() {
-    let tc = TypeChecker::new();
-    let mut refs_by_type = HashMap::from([
-        (
-            "Point".to_string(),
-            VecDeque::from(vec![BehaviorRefMetadata {
-                name: "Json".to_string(),
-                type_args: vec![AstType::I32],
-            }]),
-        ),
-        (
-            "Label".to_string(),
-            VecDeque::from(vec![BehaviorRefMetadata {
-                name: "Debug".to_string(),
-                type_args: vec![],
-            }]),
-        ),
-    ]);
-    let missing_refs = HashSet::new();
-
-    assert_eq!(
-        tc.resolver_behavior_ref_owner_for(&refs_by_type, &missing_refs, "Json", &[AstType::I32]),
-        Some("Point".to_string())
-    );
-    assert_eq!(
-        tc.resolver_behavior_ref_owner_for(&refs_by_type, &missing_refs, "Missing", &[]),
-        None
-    );
-
-    refs_by_type.remove("Label");
-    assert_eq!(
-        tc.resolver_behavior_ref_owner_for(&refs_by_type, &missing_refs, "Missing", &[]),
-        Some("Point".to_string())
-    );
-
-    refs_by_type.clear();
-    let missing_refs = HashSet::from(["Recovered".to_string()]);
-    assert_eq!(
-        tc.resolver_behavior_ref_owner_for(&refs_by_type, &missing_refs, "Missing", &[]),
-        Some("Recovered".to_string())
-    );
-}
-
-#[test]
-fn resolver_symbol_metadata_helper_requires_symbol_and_selected_metadata() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-
-    assert_eq!(
-        TypeChecker::resolver_symbol_metadata(&symbols, Namespace::Type, "Point", |symbol| {
-            symbol.field_types.as_ref()
-        })
-        .map(|(_, fields)| fields[0].0.as_str()),
-        Some("x")
-    );
-    symbols.set_field_types_for_test(Namespace::Type, "Point", None);
-    assert!(
-        TypeChecker::resolver_symbol_metadata(&symbols, Namespace::Type, "Point", |symbol| symbol
-            .field_types
-            .as_ref())
-        .is_none()
-    );
-    assert!(TypeChecker::resolver_symbol_metadata(
-        &symbols,
-        Namespace::Type,
-        "Missing",
-        |symbol| symbol.field_types.as_ref()
-    )
-    .is_none());
-}
-
-#[test]
-fn resolver_callable_signature_metadata_requires_complete_signature() {
-    let program = parse_program(
-        r#"
-identity<T> = (value: T) T { value }
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let signature = TypeChecker::resolver_callable_signature_metadata(
-        symbols
-            .lookup(Namespace::Value, "identity")
-            .expect("identity symbol"),
-    )
-    .expect("complete resolver signature");
-
-    assert_eq!(signature.parameter_names, ["value"]);
-    assert_eq!(signature.parameter_types, [AstType::Named("T".to_string())]);
-    assert_eq!(signature.return_type, &AstType::Named("T".to_string()));
-
-    symbols.set_parameter_types_for_test(Namespace::Value, "identity", None);
-    assert!(TypeChecker::resolver_callable_signature_metadata(
-        symbols
-            .lookup(Namespace::Value, "identity")
-            .expect("identity symbol")
-    )
-    .is_none());
-}
-
-#[test]
-fn resolver_struct_field_metadata_requires_field_types() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let fields = TypeChecker::resolver_struct_field_metadata(
-        symbols
-            .lookup(Namespace::Type, "Point")
-            .expect("Point symbol"),
-    )
-    .expect("complete resolver fields");
-
-    assert_eq!(fields, [("x".to_string(), AstType::I32)]);
-
-    symbols.set_field_types_for_test(Namespace::Type, "Point", None);
-    assert!(TypeChecker::resolver_struct_field_metadata(
-        symbols
-            .lookup(Namespace::Type, "Point")
-            .expect("Point symbol")
-    )
-    .is_none());
-}
-
-#[test]
-fn resolver_enum_variant_name_metadata_requires_variant_names() {
-    let program = parse_program(
-        r#"
-Option: Some(i32), None
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let variants = TypeChecker::resolver_enum_variant_name_metadata(
-        symbols
-            .lookup(Namespace::Type, "Option")
-            .expect("Option symbol"),
-    )
-    .expect("complete resolver variants");
-
-    assert_eq!(variants, ["Some", "None"]);
-
-    symbols.set_variant_names_for_test(Namespace::Type, "Option", None);
-    assert!(TypeChecker::resolver_enum_variant_name_metadata(
-        symbols
-            .lookup(Namespace::Type, "Option")
-            .expect("Option symbol")
-    )
-    .is_none());
-}
-
-#[test]
-fn resolver_behavior_method_metadata_requires_method_types() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let methods = TypeChecker::resolver_behavior_method_metadata(
-        symbols
-            .lookup(Namespace::Behavior, "Json")
-            .expect("Json symbol"),
-    )
-    .expect("complete resolver behavior methods");
-
-    assert_eq!(methods.len(), 1);
-    assert_eq!(methods[0].name, "encode");
-    assert_eq!(methods[0].parameter_types, [AstType::SelfType]);
-    assert_eq!(methods[0].return_type, AstType::Str);
-
-    symbols.set_behavior_method_types_for_test(Namespace::Behavior, "Json", None);
-    assert!(TypeChecker::resolver_behavior_method_metadata(
-        symbols
-            .lookup(Namespace::Behavior, "Json")
-            .expect("Json symbol")
-    )
-    .is_none());
-}
-
-#[test]
-fn resolver_type_parameter_metadata_requires_names_and_bound_refs() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-identity<T: Json<T>> = (value: T) T { value }
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let metadata = resolver_type_parameter_metadata(
-        symbols
-            .lookup(Namespace::Value, "identity")
-            .expect("identity symbol"),
-    )
-    .expect("complete resolver type-parameter metadata");
-
-    assert_eq!(metadata.names, ["T"]);
-    assert_eq!(metadata.bound_refs.len(), 1);
-    assert_eq!(metadata.bound_refs[0].type_parameter, "T");
-    assert_eq!(metadata.bound_refs[0].behavior, "Json");
-    assert_eq!(
-        metadata.bound_refs[0].type_args,
-        [AstType::Named("T".to_string())]
-    );
-
-    symbols.set_type_parameter_bound_refs_for_test(Namespace::Value, "identity", None);
-    assert!(resolver_type_parameter_metadata(
-        symbols
-            .lookup(Namespace::Value, "identity")
-            .expect("identity symbol")
-    )
-    .is_none());
-}
-
-#[test]
-fn callable_signature_insert_routes_function_and_method_keys() {
-    let mut tc = TypeChecker::new();
-    let function = FuncInfo {
-        name: "make".to_string(),
-        params: vec![],
-        return_type: AstType::I32,
-        type_params: vec![],
-        type_param_bounds: HashMap::new(),
-    };
-    let method = FuncInfo {
-        name: "Point.get".to_string(),
-        params: vec![("self".to_string(), AstType::Named("Point".to_string()))],
-        return_type: AstType::I32,
-        type_params: vec![],
-        type_param_bounds: HashMap::new(),
-    };
-
-    tc.insert_callable_signature("make", function);
-    tc.insert_callable_signature("Point.get", method);
-
-    assert!(tc.functions.contains_key("make"));
-    assert!(!tc.methods.contains_key("make"));
-    assert!(tc.methods.contains_key("Point.get"));
-    assert!(!tc.functions.contains_key("Point.get"));
-}
-
-#[test]
-fn generic_callable_template_mut_routes_function_and_method_keys() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T
-}
-
-identity<T> = (value: T) T { value }
-
-Box.get<T> = (self: Box<T>) T { self.value }
-"#,
-    );
-    let ast::Declaration::Function {
-        type_params: function_type_params,
-        params: function_params,
-        return_type: function_return_type,
-        body: function_body,
-        span: function_span,
-        ..
-    } = &program.declarations[1]
-    else {
-        panic!("expected generic function");
-    };
-    let ast::Declaration::Method {
-        type_params: method_type_params,
-        params: method_params,
-        return_type: method_return_type,
-        body: method_body,
-        span: method_span,
-        ..
-    } = &program.declarations[2]
-    else {
-        panic!("expected generic method");
-    };
-    let mut tc = TypeChecker::new();
-    tc.generic_functions.insert(
-        "identity".to_string(),
-        generic_template_from_type_params(
-            function_type_params,
-            function_params,
-            function_return_type,
-            function_body,
-            *function_span,
-        )
-        .expect("generic function template"),
-    );
-    tc.generic_methods.insert(
-        "Box.get".to_string(),
-        generic_template_from_type_params(
-            method_type_params,
-            method_params,
-            method_return_type,
-            method_body,
-            *method_span,
-        )
-        .expect("generic method template"),
-    );
-
-    tc.generic_callable_template_mut("identity")
-        .expect("function template")
-        .return_type = Some(AstType::I32);
-    tc.generic_callable_template_mut("Box.get")
-        .expect("method template")
-        .return_type = Some(AstType::Bool);
-
-    assert_eq!(
-        tc.generic_functions
-            .get("identity")
-            .and_then(|template| template.return_type.as_ref()),
-        Some(&AstType::I32)
-    );
-    assert_eq!(
-        tc.generic_methods
-            .get("Box.get")
-            .and_then(|template| template.return_type.as_ref()),
-        Some(&AstType::Bool)
-    );
-    assert!(!tc.generic_methods.contains_key("identity"));
-    assert!(!tc.generic_functions.contains_key("Box.get"));
-}
-
-#[test]
-fn callable_template_rekey_routes_function_and_method_keys() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T
-}
-
-identity<T> = (value: T) T { value }
-
-Box.get<T> = (self: Box<T>) T { self.value }
-"#,
-    );
-    let ast::Declaration::Function {
-        type_params: function_type_params,
-        params: function_params,
-        return_type: function_return_type,
-        body: function_body,
-        span: function_span,
-        ..
-    } = &program.declarations[1]
-    else {
-        panic!("expected generic function");
-    };
-    let ast::Declaration::Method {
-        type_params: method_type_params,
-        params: method_params,
-        return_type: method_return_type,
-        body: method_body,
-        span: method_span,
-        ..
-    } = &program.declarations[2]
-    else {
-        panic!("expected generic method");
-    };
-    let mut tc = TypeChecker::new();
-    tc.generic_functions.insert(
-        "identity".to_string(),
-        generic_template_from_type_params(
-            function_type_params,
-            function_params,
-            function_return_type,
-            function_body,
-            *function_span,
-        )
-        .expect("generic function template"),
-    );
-    tc.generic_methods.insert(
-        "Box.get".to_string(),
-        generic_template_from_type_params(
-            method_type_params,
-            method_params,
-            method_return_type,
-            method_body,
-            *method_span,
-        )
-        .expect("generic method template"),
-    );
-
-    tc.rekey_callable_template("identity", "renamed");
-    tc.rekey_callable_template("Box.get", "Box.fetch");
-
-    assert!(tc.generic_functions.contains_key("renamed"));
-    assert!(!tc.generic_functions.contains_key("identity"));
-    assert!(tc.generic_methods.contains_key("Box.fetch"));
-    assert!(!tc.generic_methods.contains_key("Box.get"));
-}
-
-#[test]
-fn resolver_backed_callable_template_collection_defers_signature_metadata_to_resolver() {
-    let program = parse_program(
-        r#"
-Box<T>: {
-    value: T
-}
-
-identity<T> = (mut value: T) T { value }
-
-Box.get<T> = (self: Box<T>, mut fallback: T) T { fallback }
-"#,
-    );
-    let mut tc = TypeChecker::new();
-
-    tc.with_resolver_backed_collection(|checker| {
-        checker.collect_declarations(&program.declarations);
-    });
-
-    let function_template = tc
-        .generic_functions
-        .get("identity")
-        .expect("function template stub");
-    assert!(
-            function_template.type_params.is_empty(),
-            "resolver-backed generic function templates should not keep AST generic names before resolver metadata"
-        );
-    assert_eq!(function_template.params.len(), 1);
-    assert_eq!(function_template.params[0].name, "");
-    assert_eq!(function_template.params[0].ty, AstType::Void);
-    assert!(function_template.params[0].mutable);
-    assert_eq!(function_template.return_type, None);
-
-    let method_template = tc
-        .generic_methods
-        .get("Box.get")
-        .expect("method template stub");
-    assert!(
-            method_template.type_params.is_empty(),
-            "resolver-backed generic method templates should not keep AST generic names before resolver metadata"
-        );
-    assert_eq!(method_template.params.len(), 2);
-    assert_eq!(method_template.params[1].name, "");
-    assert_eq!(method_template.params[1].ty, AstType::Void);
-    assert!(method_template.params[1].mutable);
-    assert_eq!(method_template.return_type, None);
-}
-
-#[test]
-fn behavior_ref_validation_maps_role_and_check_diagnostics() {
-    let cases = [
-        (
-            BehaviorRefRole::Parent,
-            BehaviorRefCheck::Contains,
-            ("behavior", "parents", "parent refs", "E0235", "E0245"),
-        ),
-        (
-            BehaviorRefRole::Parent,
-            BehaviorRefCheck::List,
-            ("behavior", "parents", "parent refs", "E0240", "E0246"),
-        ),
-        (
-            BehaviorRefRole::Impl,
-            BehaviorRefCheck::Contains,
-            (
-                "type",
-                "behavior impls",
-                "behavior impl refs",
-                "E0236",
-                "E0247",
-            ),
-        ),
-        (
-            BehaviorRefRole::Impl,
-            BehaviorRefCheck::List,
-            (
-                "type",
-                "behavior impls",
-                "behavior impl refs",
-                "E0238",
-                "E0248",
-            ),
-        ),
-        (
-            BehaviorRefRole::Required,
-            BehaviorRefCheck::Contains,
-            (
-                "type",
-                "behavior requires",
-                "behavior requires refs",
-                "E0237",
-                "E0249",
-            ),
-        ),
-        (
-            BehaviorRefRole::Required,
-            BehaviorRefCheck::List,
-            (
-                "type",
-                "behavior requires",
-                "behavior requires refs",
-                "E0239",
-                "E0250",
-            ),
-        ),
-    ];
-
-    for (role, check, expected) in cases {
-        let validation = BehaviorRefValidation::for_role(role, check);
-        assert_eq!(
-            (
-                validation.symbol_kind,
-                validation.name_label,
-                validation.ref_label,
-                validation.name_code,
-                validation.ref_code,
-            ),
-            expected
-        );
-    }
-
-    let contains =
-        BehaviorRefValidation::for_role(BehaviorRefRole::Impl, BehaviorRefCheck::Contains);
-    assert_eq!(
-            contains.contains_name_message("Point", "PrettyJson", "Json<str>"),
-            "resolver type symbol 'Point' has behavior impls 'PrettyJson', expected to include 'Json<str>'"
-        );
-    assert_eq!(
-            contains.contains_ref_message("Point", "PrettyJson", "Json<str>"),
-            "resolver type symbol 'Point' has behavior impl refs 'PrettyJson', expected to include 'Json<str>'"
-        );
-
-    let list = BehaviorRefValidation::for_role(BehaviorRefRole::Parent, BehaviorRefCheck::List);
-    assert_eq!(
-        list.list_name_message("PrettyJson", "Json, Debug", "Json"),
-        "resolver behavior symbol 'PrettyJson' has parents 'Json, Debug', expected 'Json'"
-    );
-    assert_eq!(
-        list.list_ref_message("PrettyJson", "Json, Debug", "Json"),
-        "resolver behavior symbol 'PrettyJson' has parent refs 'Json, Debug', expected 'Json'"
-    );
-}
-
-#[test]
-fn behavior_ref_validation_separates_role_labels_from_check_codes() {
-    let parent = BehaviorRefValidation::role_labels(BehaviorRefRole::Parent);
-    let implementation = BehaviorRefValidation::role_labels(BehaviorRefRole::Impl);
-    let required = BehaviorRefValidation::role_labels(BehaviorRefRole::Required);
-    let parent_contains =
-        BehaviorRefValidation::codes_for(BehaviorRefRole::Parent, BehaviorRefCheck::Contains);
-    let parent_list =
-        BehaviorRefValidation::codes_for(BehaviorRefRole::Parent, BehaviorRefCheck::List);
-
-    assert_eq!(parent, ("behavior", "parents", "parent refs"));
-    assert_eq!(
-        implementation,
-        ("type", "behavior impls", "behavior impl refs")
-    );
-    assert_eq!(
-        required,
-        ("type", "behavior requires", "behavior requires refs")
-    );
-    assert_eq!(parent_contains, ("E0235", "E0245"));
-    assert_eq!(parent_list, ("E0240", "E0246"));
-}
-
-#[test]
-fn behavior_ref_actual_selects_role_metadata() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(PrettyJson) {
-    encode = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let behavior = symbols
-        .lookup(Namespace::Behavior, "PrettyJson")
-        .expect("behavior symbol");
-    let ty = symbols
-        .lookup(Namespace::Type, "Point")
-        .expect("type symbol");
-
-    let parent = BehaviorRefActual::for_role(behavior, BehaviorRefRole::Parent);
-    assert_eq!(format_behavior_ref_names(parent.names), "Json<str>");
-    assert_eq!(format_behavior_refs(parent.refs), "Json<str>");
-
-    let implementation = BehaviorRefActual::for_role(ty, BehaviorRefRole::Impl);
-    assert_eq!(
-        format_behavior_ref_names(implementation.names),
-        "PrettyJson"
-    );
-    assert_eq!(format_behavior_refs(implementation.refs), "PrettyJson");
-
-    let required = BehaviorRefActual::for_role(ty, BehaviorRefRole::Required);
-    assert_eq!(format_behavior_ref_names(required.names), "Json<str>");
-    assert_eq!(format_behavior_refs(required.refs), "Json<str>");
-}
-
-#[test]
-fn behavior_ref_actual_exposes_role_metadata_selection() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(PrettyJson) {
-    encode = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let behavior = symbols
-        .lookup(Namespace::Behavior, "PrettyJson")
-        .expect("behavior symbol");
-    let ty = symbols
-        .lookup(Namespace::Type, "Point")
-        .expect("type symbol");
-
-    let (parent_names, parent_refs) =
-        BehaviorRefActual::metadata_for_role(behavior, BehaviorRefRole::Parent);
-    let (impl_names, impl_refs) = BehaviorRefActual::metadata_for_role(ty, BehaviorRefRole::Impl);
-    let (required_names, required_refs) =
-        BehaviorRefActual::metadata_for_role(ty, BehaviorRefRole::Required);
-
-    assert_eq!(format_behavior_ref_names(parent_names), "Json<str>");
-    assert_eq!(format_behavior_refs(parent_refs), "Json<str>");
-    assert_eq!(format_behavior_ref_names(impl_names), "PrettyJson");
-    assert_eq!(format_behavior_refs(impl_refs), "PrettyJson");
-    assert_eq!(format_behavior_ref_names(required_names), "Json<str>");
-    assert_eq!(format_behavior_refs(required_refs), "Json<str>");
-}
-
-#[test]
-fn behavior_ref_actual_matches_expected_edges() {
-    let names = vec!["Json<i32>".to_string()];
-    let refs = vec![BehaviorRefMetadata {
-        name: "Json".to_string(),
-        type_args: vec![AstType::I32],
-    }];
-    let actual = BehaviorRefActual {
-        names: Some(&names),
-        refs: Some(&refs),
-    };
-    let expected = expected_behavior_edge("Json", &[AstType::I32]);
-    let mismatch = expected_behavior_edge("Debug", &[]);
-    let expected_list = ExpectedBehaviorEdgeMetadata::from_edges(std::slice::from_ref(&expected));
-
-    assert!(actual.contains_display(&expected.display));
-    assert!(actual.contains_metadata(&expected.metadata));
-    assert!(!actual.contains_display(&mismatch.display));
-    assert!(!actual.contains_metadata(&mismatch.metadata));
-    assert!(actual.names_match(&expected_list.names));
-    assert!(actual.refs_match(&expected_list.refs));
-}
-
-#[test]
-fn expected_parameter_builds_name_display_and_type_together() {
-    let parameter = ExpectedParameter::new(
-        "mapper",
-        &AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        },
-    );
-
-    assert_eq!(parameter.name, "mapper");
-    assert_eq!(parameter.display, "(i32) str");
-    assert_eq!(
-        parameter.typed,
-        AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        }
-    );
-}
-
-#[test]
-fn expected_return_metadata_defaults_and_displays_together() {
-    let explicit = ExpectedReturnMetadata::new(&Some(AstType::Named("Point".to_string())));
-    let implicit = ExpectedReturnMetadata::new(&None);
-
-    assert_eq!(explicit.display, "Point");
-    assert_eq!(explicit.typed, AstType::Named("Point".to_string()));
-    assert_eq!(implicit.display, "void");
-    assert_eq!(implicit.typed, AstType::Void);
-}
-
-#[test]
-fn expected_type_parameter_builds_bound_display_and_ref_together() {
-    let type_param = ast::TypeParam {
-        name: "T".to_string(),
-        constraint: Some("Json".to_string()),
-        constraint_type_args: vec![AstType::Named("T".to_string())],
-        span: Span::dummy(),
-    };
-
-    let expected = ExpectedTypeParameter::new(&type_param);
-    let bound = expected.bound.expect("expected bound");
-
-    assert_eq!(expected.name, "T");
-    assert_eq!(bound.display, ("T".to_string(), "Json<T>".to_string()));
-    assert_eq!(bound.reference.type_parameter, "T");
-    assert_eq!(bound.reference.behavior, "Json");
-    assert_eq!(
-        bound.reference.type_args,
-        vec![AstType::Named("T".to_string())]
-    );
-}
-
-#[test]
-fn expected_field_builds_display_and_type_together() {
-    let field = ExpectedField::new(
-        "mapper",
-        &AstType::Function {
-            params: vec![AstType::Named("Input".to_string())],
-            ret: Box::new(AstType::Named("Output".to_string())),
-        },
-    );
-
-    assert_eq!(
-        field.display,
-        ("mapper".to_string(), "(Input) Output".to_string())
-    );
-    assert_eq!(
-        field.typed,
-        (
-            "mapper".to_string(),
-            AstType::Function {
-                params: vec![AstType::Named("Input".to_string())],
-                ret: Box::new(AstType::Named("Output".to_string())),
-            }
-        )
-    );
-}
-
-#[test]
-fn expected_variant_payload_builds_display_and_type_together() {
-    let payload = ExpectedVariantPayloadType::new(&Some(AstType::Function {
-        params: vec![AstType::I32],
-        ret: Box::new(AstType::Bool),
-    }));
-    let empty_payload = ExpectedVariantPayloadType::new(&None);
-
-    assert_eq!(payload.display, Some("(i32) bool".to_string()));
-    assert_eq!(
-        payload.typed,
-        Some(AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Bool),
-        })
-    );
-    assert_eq!(empty_payload.display, None);
-    assert_eq!(empty_payload.typed, None);
-}
-
-#[test]
-fn expected_behavior_method_builds_signature_and_metadata_together() {
-    let method = ast::BehaviorMethod {
-        name: "map".to_string(),
-        params: vec![Param {
-            name: "mapper".to_string(),
-            ty: AstType::Function {
-                params: vec![AstType::I32],
-                ret: Box::new(AstType::Str),
-            },
-            mutable: false,
-            span: Span::dummy(),
-        }],
-        return_type: Some(AstType::Str),
-        default_body: None,
-        span: Span::dummy(),
-    };
-
-    let expected = ExpectedBehaviorMethod::new(&method);
-
-    assert_eq!(
-        expected.signature,
-        (
-            "map".to_string(),
-            vec!["(i32) str".to_string()],
-            "str".to_string(),
-        )
-    );
-    assert_eq!(expected.metadata.name, "map");
-    assert_eq!(
-        expected.metadata.parameter_names,
-        vec!["mapper".to_string()]
-    );
-    assert_eq!(
-        expected.metadata.parameter_types,
-        vec![AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Str),
-        }]
-    );
-    assert_eq!(expected.metadata.return_type, AstType::Str);
-}
-
-#[test]
-fn expected_value_signature_builds_components_together() {
-    let params = vec![Param {
-        name: "value".to_string(),
-        ty: AstType::Named("T".to_string()),
-        mutable: false,
-        span: Span::dummy(),
-    }];
-    let return_type = Some(AstType::Named("T".to_string()));
-    let type_params = vec![ast::TypeParam {
-        name: "T".to_string(),
-        constraint: Some("Json".to_string()),
-        constraint_type_args: vec![AstType::Named("T".to_string())],
-        span: Span::dummy(),
-    }];
-
-    let signature = ExpectedValueSignature::new(&params, &return_type, &type_params);
-
-    assert_eq!(signature.params[0].name, "value");
-    assert_eq!(signature.params[0].display, "T");
-    assert_eq!(signature.params[0].typed, AstType::Named("T".to_string()));
-    assert_eq!(signature.return_type.display, "T");
-    assert_eq!(signature.return_type.typed, AstType::Named("T".to_string()));
-    assert_eq!(signature.type_params[0].name, "T");
-    let bound = signature.type_params[0]
-        .bound
-        .as_ref()
-        .expect("expected bound");
-    assert_eq!(bound.display, ("T".to_string(), "Json<T>".to_string()));
-    assert_eq!(bound.reference.behavior, "Json");
-    assert_eq!(
-        bound.reference.type_args,
-        vec![AstType::Named("T".to_string())]
-    );
-}
-
-#[test]
-fn expected_value_symbol_builds_signature_and_visibility_together() {
-    let params = vec![Param {
-        name: "value".to_string(),
-        ty: AstType::I32,
-        mutable: false,
-        span: Span::dummy(),
-    }];
-    let return_type = Some(AstType::Bool);
-
-    let symbol = ExpectedValueSymbol::new(&params, &return_type, &[], true);
-
-    assert!(symbol.is_public);
-    assert_eq!(symbol.signature.params[0].name, "value");
-    assert_eq!(symbol.signature.params[0].display, "i32");
-    assert_eq!(symbol.signature.params[0].typed, AstType::I32);
-    assert_eq!(symbol.signature.return_type.display, "bool");
-    assert_eq!(symbol.signature.return_type.typed, AstType::Bool);
-    assert!(symbol.signature.type_params.is_empty());
-}
-
-#[test]
-fn expected_type_like_symbol_builds_type_params_and_visibility_together() {
-    let type_params = vec![ast::TypeParam {
-        name: "T".to_string(),
-        constraint: Some("Json".to_string()),
-        constraint_type_args: vec![AstType::Named("T".to_string())],
-        span: Span::dummy(),
-    }];
-
-    let symbol = ExpectedTypeLikeSymbol::new(&type_params, Some(true));
-
-    assert_eq!(symbol.is_public, Some(true));
-    assert_eq!(symbol.type_params[0].name, "T");
-    let bound = symbol.type_params[0]
-        .bound
-        .as_ref()
-        .expect("expected bound");
-    assert_eq!(bound.display, ("T".to_string(), "Json<T>".to_string()));
-    assert_eq!(bound.reference.type_parameter, "T");
-    assert_eq!(bound.reference.behavior, "Json");
-    assert_eq!(
-        bound.reference.type_args,
-        vec![AstType::Named("T".to_string())]
-    );
-}
-
-#[test]
-fn expected_behavior_symbol_builds_type_like_and_methods_together() {
-    let type_params = vec![ast::TypeParam {
-        name: "T".to_string(),
-        constraint: None,
-        constraint_type_args: vec![],
-        span: Span::dummy(),
-    }];
-    let methods = vec![ast::BehaviorMethod {
-        name: "encode".to_string(),
-        params: vec![Param {
-            name: "value".to_string(),
-            ty: AstType::Named("Self".to_string()),
-            mutable: false,
-            span: Span::dummy(),
-        }],
-        return_type: Some(AstType::Named("T".to_string())),
-        default_body: None,
-        span: Span::dummy(),
-    }];
-
-    let symbol = ExpectedBehaviorSymbol::new(&type_params, &methods, true);
-
-    assert_eq!(symbol.type_like.is_public, Some(true));
-    assert_eq!(symbol.type_like.type_params[0].name, "T");
-    assert_eq!(symbol.methods[0].signature.0, "encode");
-    assert_eq!(symbol.methods[0].signature.1, vec!["Self".to_string()]);
-    assert_eq!(symbol.methods[0].signature.2, "T");
-    assert_eq!(symbol.methods[0].metadata.name, "encode");
-    assert_eq!(
-        symbol.methods[0].metadata.parameter_names,
-        vec!["value".to_string()]
-    );
-    assert_eq!(
-        symbol.methods[0].metadata.return_type,
-        AstType::Named("T".to_string())
-    );
-}
-
-#[test]
-fn expected_struct_symbol_builds_type_like_and_fields_together() {
-    let type_params = vec![ast::TypeParam {
-        name: "T".to_string(),
-        constraint: None,
-        constraint_type_args: vec![],
-        span: Span::dummy(),
-    }];
-    let fields = vec![StructField {
-        name: "value".to_string(),
-        ty: AstType::Named("T".to_string()),
-        default: None,
-        mutable: false,
-        span: Span::dummy(),
-    }];
-
-    let symbol = ExpectedStructSymbol::new(&type_params, &fields, true);
-
-    assert_eq!(symbol.type_like.is_public, Some(true));
-    assert_eq!(symbol.type_like.type_params[0].name, "T");
-    assert_eq!(
-        symbol.fields[0].display,
-        ("value".to_string(), "T".to_string())
-    );
-    assert_eq!(
-        symbol.fields[0].typed,
-        ("value".to_string(), AstType::Named("T".to_string()))
-    );
-}
-
-#[test]
-fn expected_enum_symbol_builds_type_like_and_variants_together() {
-    let type_params = vec![ast::TypeParam {
-        name: "T".to_string(),
-        constraint: None,
-        constraint_type_args: vec![],
-        span: Span::dummy(),
-    }];
-    let variants = vec![
-        EnumVariant {
-            name: "Some".to_string(),
-            payload: Some(AstType::Named("T".to_string())),
-            span: Span::dummy(),
-        },
-        EnumVariant {
-            name: "None".to_string(),
-            payload: None,
-            span: Span::dummy(),
-        },
-    ];
-
-    let symbol = ExpectedEnumSymbol::new(&type_params, &variants, true);
-
-    assert_eq!(symbol.type_like.is_public, Some(true));
-    assert_eq!(symbol.type_like.type_params[0].name, "T");
-    assert_eq!(
-        symbol.variant_names,
-        vec!["Some".to_string(), "None".to_string()]
-    );
-}
-
-#[test]
-fn expected_variant_symbol_builds_owner_visibility_and_payload_together() {
-    let payload = Some(AstType::Function {
-        params: vec![AstType::I32],
-        ret: Box::new(AstType::Bool),
-    });
-
-    let symbol = ExpectedVariantSymbol::new("Result", true, &payload);
-
-    assert_eq!(symbol.owner_name, "Result");
-    assert!(symbol.is_public);
-    assert_eq!(symbol.payload.display, Some("(i32) bool".to_string()));
-    assert_eq!(
-        symbol.payload.typed,
-        Some(AstType::Function {
-            params: vec![AstType::I32],
-            ret: Box::new(AstType::Bool),
-        })
-    );
-}
-
-#[test]
-fn expected_import_symbol_builds_source_and_visibility_together() {
-    let symbol = ExpectedImportSymbol::new("std.io");
-
-    assert_eq!(symbol.source, "std.io");
-    assert!(!symbol.is_public);
-}
-
-#[test]
-fn expected_module_symbol_builds_name_source_and_visibility_together() {
-    let symbol = ExpectedModuleSymbol::new("std.io");
-
-    assert_eq!(symbol.name, "std.io");
-    assert_eq!(symbol.source, None);
-    assert!(!symbol.is_public);
-}
-
-#[test]
-fn expected_local_symbol_builds_scope_mutability_source_and_visibility_together() {
-    let symbol = ExpectedLocalSymbol::new(true, 42);
-
-    assert_eq!(symbol.scope_id, 42);
-    assert!(symbol.is_mutable);
-    assert_eq!(symbol.source, None);
-    assert!(!symbol.is_public);
-}
-
-#[test]
-fn expected_behavior_edge_builds_display_and_metadata_together() {
-    let edge = ExpectedBehaviorEdge::new("Json", &[AstType::I32]);
-
-    assert_eq!(edge.display, "Json<i32>");
-    assert_eq!(edge.metadata.name, "Json");
-    assert_eq!(edge.metadata.type_args, vec![AstType::I32]);
-}
-
-#[test]
-fn expected_behavior_associations_build_impl_and_required_edges_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-
-    let expected = ExpectedBehaviorAssociations::new(&program);
-    let impl_edge = &expected.impls.edges_for("Point")[0];
-    let required_edge = &expected.required.edges_for("Point")[0];
-
-    assert_eq!(impl_edge.display, "Json<str>");
-    assert_eq!(impl_edge.metadata.name, "Json");
-    assert_eq!(impl_edge.metadata.type_args, vec![AstType::Str]);
-    assert_eq!(required_edge.display, "Json<str>");
-    assert_eq!(required_edge.metadata.name, "Json");
-    assert_eq!(required_edge.metadata.type_args, vec![AstType::Str]);
-}
-
-#[test]
-fn resolver_behavior_association_list_tasks_collect_type_and_parent_edges_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-
-    let tasks = TypeChecker::collect_resolver_behavior_association_list_tasks(&program, &symbols);
-
-    assert_eq!(tasks.type_associations.len(), 1);
-    let type_task = &tasks.type_associations[0];
-    assert_eq!(type_task.name, "Point");
-    assert_eq!(type_task.impl_edges[0].display, "Json<str>");
-    assert_eq!(type_task.required_edges[0].display, "Json<str>");
-
-    assert_eq!(tasks.behavior_parents.len(), 2);
-    let pretty_task = tasks
-        .behavior_parents
-        .iter()
-        .find(|task| task.name == "PrettyJson")
-        .expect("PrettyJson parent task");
-    assert_eq!(pretty_task.parent_edges[0].display, "Json<str>");
-    let json_task = tasks
-        .behavior_parents
-        .iter()
-        .find(|task| task.name == "Json")
-        .expect("Json empty parent task");
-    assert!(json_task.parent_edges.is_empty());
-}
-
-#[test]
-fn resolver_expected_symbol_sets_collect_declarations_and_locals_together() {
-    let program = parse_program(
-        r#"
-{ io } = std
-
-Point: { x: i32 = 1 }
-
-main = (input: i32) i32 {
-    value := input
-    value
-}
-"#,
-    );
-
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let expected =
-        TypeChecker::collect_resolver_validation_replay_tasks(&program, &symbols).expected_symbols;
-
-    assert!(expected.validate_imports);
-    assert!(expected
-        .declarations
-        .contains(&(Namespace::Module, "std".to_string())));
-    assert!(expected
-        .declarations
-        .contains(&(Namespace::Import, "io".to_string())));
-    assert!(expected
-        .declarations
-        .contains(&(Namespace::Type, "Point".to_string())));
-    assert!(expected
-        .declarations
-        .contains(&(Namespace::Value, "main".to_string())));
-    assert!(expected.locals.contains(&("input".to_string(), 2)));
-    assert!(expected.locals.contains(&("value".to_string(), 3)));
-}
-
-#[test]
-fn resolver_validation_replay_declaration_tasks_collect_sources_and_edges() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-
-    let tasks =
-        TypeChecker::collect_resolver_validation_replay_declaration_tasks(&program, &symbols);
-
-    assert!(tasks
-        .expected_symbols
-        .declarations
-        .contains(&(Namespace::Type, "Point".to_string())));
-    assert_eq!(tasks.type_declarations.len(), 1);
-    assert_eq!(tasks.type_declarations[0].name, "Point");
-    assert_eq!(
-        tasks.expected_associations.impls.owned_edges_for("Point")[0].display,
-        "Json<str>"
-    );
-    assert_eq!(
-        tasks
-            .expected_associations
-            .required
-            .owned_edges_for("Point")[0]
-            .display,
-        "Json<str>"
-    );
-    assert_eq!(tasks.behavior_declarations.len(), 1);
-    assert_eq!(tasks.behavior_declarations[0].name, "Json");
-}
-
-#[test]
-fn resolver_type_reference_validation_tasks_collect_only_type_reference_work() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-main = (input: Point) Point {
-    input
-}
-"#,
-    );
-
-    let tasks =
-        TypeChecker::collect_resolver_type_reference_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    assert!(matches!(
-        tasks[0],
-        ResolverTypeReferenceValidationTask::Struct { name: "Point", .. }
-    ));
-    assert!(matches!(
-        tasks[1],
-        ResolverTypeReferenceValidationTask::Function { name: "main", .. }
-    ));
-}
-
-#[test]
-fn resolver_validation_replay_tasks_collect_symbols_and_behavior_associations_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-
-main = (input: i32) i32 {
-    input
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-
-    let tasks = TypeChecker::collect_resolver_validation_replay_tasks(&program, &symbols);
-
-    assert!(tasks
-        .expected_symbols
-        .declarations
-        .contains(&(Namespace::Type, "Point".to_string())));
-    assert!(tasks
-        .expected_symbols
-        .declarations
-        .contains(&(Namespace::Behavior, "Json".to_string())));
-    assert!(tasks
-        .expected_symbols
-        .declarations
-        .contains(&(Namespace::Value, "main".to_string())));
-    assert!(tasks
-        .expected_symbols
-        .locals
-        .iter()
-        .any(|(name, _)| name == "input"));
-
-    let type_task = &tasks.behavior_associations.type_associations[0];
-    assert_eq!(type_task.name, "Point");
-    assert_eq!(type_task.impl_edges[0].display, "Json<str>");
-    assert_eq!(type_task.required_edges[0].display, "Json<str>");
-    assert_eq!(tasks.behavior_associations.behavior_parents.len(), 1);
-}
-
-#[test]
-fn resolver_declaration_metadata_tasks_collect_top_level_type_reference_tasks() {
-    let program = parse_program(
-        r#"
-value := 1
-"#,
-    );
-
-    let tasks = TypeChecker::collect_resolver_declaration_metadata_tasks(&program.declarations);
-
-    assert!(
-        tasks.type_references.iter().any(|task| matches!(
-            task,
-            ResolverTypeReferenceValidationTask::TopLevelExpr { .. }
-        )),
-        "top-level expression type references should stay in the shared resolver task collector"
-    );
-}
-
-#[test]
-fn behavior_extends_replay_task_helper_pushes_parent_validation() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-}
-Pretty<T>: behavior {
-}
-
-Pretty.extends(Json<T>)
-"#,
-    );
-    let mut tasks = Vec::new();
-
-    let handled =
-        TypeChecker::push_behavior_extends_replay_task(&program.declarations[2], &mut tasks);
-
-    assert!(handled);
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].behavior, "Pretty");
-    assert_eq!(tasks[0].parent, "Json");
-    assert_eq!(
-        tasks[0].parent_type_args,
-        &[AstType::Named("T".to_string())]
-    );
-}
-
-#[test]
-fn behavior_requires_replay_task_helper_pushes_requires_validation() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let mut tasks = Vec::new();
-
-    let handled =
-        TypeChecker::push_behavior_requires_replay_task(&program.declarations[2], &mut tasks);
-
-    assert!(handled);
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].type_name, "Point");
-    assert_eq!(tasks[0].behavior, "Json");
-    assert_eq!(tasks[0].behavior_type_args, &[AstType::Str]);
-}
-
-#[test]
-fn behavior_association_validation_tasks_collect_extends_impls_and_requires_together() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Pretty<T>: behavior {
-    pretty: (Self) T
-}
-
-Pretty.extends(Json<T>)
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-
-    let tasks = TypeChecker::collect_behavior_association_validation_tasks(&program.declarations);
-
-    assert_eq!(tasks.extends.len(), 1);
-    assert_eq!(tasks.extends[0].behavior, "Pretty");
-    assert_eq!(tasks.extends[0].parent, "Json");
-    assert_eq!(
-        tasks.extends[0].parent_type_args,
-        &[AstType::Named("T".to_string())]
-    );
-    assert_eq!(tasks.impls.len(), 1);
-    assert_eq!(tasks.impls[0].ast_type_name, "Point");
-    assert_eq!(tasks.impls[0].behavior, "Json");
-    assert_eq!(tasks.impls[0].behavior_type_args, &[AstType::Str]);
-    assert_eq!(tasks.requires.len(), 1);
-    assert_eq!(tasks.requires[0].type_name, "Point");
-    assert_eq!(tasks.requires[0].behavior, "Json");
-    assert_eq!(tasks.requires[0].behavior_type_args, &[AstType::Str]);
-}
-
-#[test]
-fn behavior_association_validation_helper_replays_impls_and_requires() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-    let mut checker = TypeChecker::new();
-    checker.collect_declarations(&program.declarations);
-    let tasks = TypeChecker::collect_behavior_association_validation_tasks(&program.declarations);
-
-    checker.validate_behavior_association_tasks(&tasks, None);
-
-    assert!(
-        checker.diagnostics().is_empty(),
-        "valid impl+requires replay should not emit diagnostics: {:?}",
-        checker.diagnostics()
-    );
-}
-
-#[test]
-fn impl_block_declaration_tasks_collect_behavior_and_plain_impls() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.impl = {
-    x_value = (value: Point) i32 { value.x }
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let tasks = TypeChecker::collect_impl_block_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    assert_eq!(tasks[0].type_name, "Point");
-    assert_eq!(tasks[0].behavior, None);
-    assert_eq!(tasks[1].type_name, "Point");
-    assert_eq!(tasks[1].behavior, Some("Json"));
-    assert_eq!(tasks[1].methods.len(), 1);
-}
-
-#[test]
-fn callable_declaration_tasks_collect_functions_and_methods() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-make = () Point { Point { x: 1 } }
-
-Point.get = (self: Point) i32 { self.x }
-"#,
-    );
-
-    let tasks = TypeChecker::collect_callable_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    match &tasks[0] {
-        CallableDeclarationTask::Function { name, .. } => assert_eq!(*name, "make"),
-        _ => panic!("expected function task"),
-    }
-    match &tasks[1] {
-        CallableDeclarationTask::Method {
-            type_name,
-            method_name,
-            ..
-        } => {
-            assert_eq!(*type_name, "Point");
-            assert_eq!(*method_name, "get");
-        }
-        _ => panic!("expected method task"),
-    }
-}
-
-#[test]
-fn ast_type_declaration_tasks_collect_structs_and_enums() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Option<T>: Some(T), None
-"#,
-    );
-
-    let tasks = TypeChecker::collect_ast_type_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 2);
-    match &tasks[0] {
-        AstTypeDeclarationTask::Struct { name, fields, .. } => {
-            assert_eq!(*name, "Point");
-            assert_eq!(fields.len(), 1);
-        }
-        _ => panic!("expected struct task"),
-    }
-    match &tasks[1] {
-        AstTypeDeclarationTask::Enum {
-            name,
-            type_params,
-            variants,
-        } => {
-            assert_eq!(*name, "Option");
-            assert_eq!(type_params.len(), 1);
-            assert_eq!(variants.len(), 2);
-        }
-        _ => panic!("expected enum task"),
-    }
-}
-
-#[test]
-fn behavior_declaration_tasks_collect_behavior_signatures() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-"#,
-    );
-
-    let tasks = TypeChecker::collect_behavior_declaration_tasks(&program.declarations);
-
-    assert_eq!(tasks.len(), 1);
-    assert_eq!(tasks[0].name, "Json");
-    assert_eq!(tasks[0].type_params.len(), 1);
-    assert_eq!(tasks[0].methods.len(), 1);
-    assert_eq!(tasks[0].methods[0].name, "encode");
-}
-
-#[test]
-fn expected_resolver_impl_method_symbols_collect_value_symbols_and_locals() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Point.impl = {
-    x_value = (value: Point) i32 { value.x }
-}
-"#,
-    );
-    let Declaration::ImplBlock {
-        type_name, methods, ..
-    } = &program.declarations[1]
-    else {
-        panic!("expected impl block");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let mut expected = ResolverExpectedSymbolSets::default();
-
-    collect_expected_resolver_impl_method_symbols(
-        type_name,
-        methods,
-        &mut scope_cursor,
-        &mut expected,
-    );
-
-    assert!(expected
-        .declarations
-        .contains(&(Namespace::Value, "Point.x_value".to_string())));
-    assert!(expected.locals.iter().any(|(name, _)| name == "value"));
-}
-
-#[test]
-fn expected_resolver_callable_locals_collect_params_and_body() {
-    let program = parse_program(
-        r#"
-main = (input: i32) i32 {
-    value := input
-    value
-}
-"#,
-    );
-    let Declaration::Function { params, body, .. } = &program.declarations[0] else {
-        panic!("expected function");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let mut expected = HashSet::new();
-
-    expected_resolver_callable_locals(params, body, &mut scope_cursor, &mut expected);
-
-    assert!(expected.iter().any(|(name, _)| name == "input"));
-    assert!(expected.iter().any(|(name, _)| name == "value"));
-}
-
-#[test]
-fn expected_resolver_scoped_expr_locals_collects_block_bindings() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    value := 1
-    value
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[0] else {
-        panic!("expected function");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let mut expected = HashSet::new();
-
-    expected_resolver_scoped_expr_locals(body, &mut scope_cursor, &mut expected);
-
-    assert!(expected.iter().any(|(name, _)| name == "value"));
-}
-
-#[test]
-fn expected_resolver_child_expr_locals_collects_branch_bindings() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    loop {
-        value := 1
-        break
-    }
-    value
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[0] else {
-        panic!("expected function");
-    };
-    let Expression::Block { statements, .. } = body else {
-        panic!("expected block");
-    };
-    let Some(ast::Statement::Expression { expr, .. }) = statements.first() else {
-        panic!("expected expression statement");
-    };
-    let Expression::Loop { body, .. } = expr else {
-        panic!("expected loop expression");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let locals = scope_cursor.new_scope();
-    let mut expected = HashSet::new();
-
-    expected_resolver_child_expr_locals(body, &mut scope_cursor, &locals, &mut expected);
-
-    assert!(expected.iter().any(|(name, _)| name == "value"));
-}
-
-#[test]
-fn expected_resolver_pattern_expr_locals_collects_pattern_and_body_bindings() {
-    let program = parse_program(
-        r#"
-Option:
-    None,
-    Some(i32)
-
-main = (value: Option) i32 {
-    value ?
-        | Some(inner) {
-            doubled := inner
-            doubled
-        }
-        | None { 0 }
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[1] else {
-        panic!("expected function");
-    };
-    let Expression::Block {
-        expr: Some(expr), ..
-    } = body
-    else {
-        panic!("expected block");
-    };
-    let Expression::Match { arms, .. } = expr.as_ref() else {
-        panic!("expected match expression");
-    };
-    let arm = arms.first().expect("first match arm");
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let locals = scope_cursor.new_scope();
-    let mut expected = HashSet::new();
-
-    expected_resolver_pattern_expr_locals(
-        &arm.pattern,
-        &arm.body,
-        &mut scope_cursor,
-        &locals,
-        &mut expected,
-    );
-
-    assert!(expected.iter().any(|(name, _)| name == "inner"));
-    assert!(expected.iter().any(|(name, _)| name == "doubled"));
-}
-
-#[test]
-fn expected_resolver_pattern_locals_collects_struct_shorthand_bindings() {
-    let program = parse_program(
-        r#"
-Point: { x: i32, y: i32 }
-
-main = (point: Point) i32 {
-    point ?
-        | Point { x, y } { x + y }
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[1] else {
-        panic!("expected function");
-    };
-    let Expression::Block {
-        expr: Some(expr), ..
-    } = body
-    else {
-        panic!("expected block");
-    };
-    let Expression::Match { arms, .. } = expr.as_ref() else {
-        panic!("expected match expression");
-    };
-    let arm = arms.first().expect("first match arm");
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let mut locals = scope_cursor.new_scope();
-    let mut expected = HashSet::new();
-
-    expected_resolver_pattern_locals(&arm.pattern, &mut scope_cursor, &mut locals, &mut expected);
-
-    assert!(expected.iter().any(|(name, _)| name == "x"));
-    assert!(expected.iter().any(|(name, _)| name == "y"));
-}
-
-#[test]
-fn expected_resolver_block_locals_collects_statement_and_final_expr_bindings() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    value := 1
-    (input: i32) i32 {
-        input
-    }
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[0] else {
-        panic!("expected function");
-    };
-    let Expression::Block {
-        statements, expr, ..
-    } = body
-    else {
-        panic!("expected block");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let locals = scope_cursor.new_scope();
-    let mut expected = HashSet::new();
-
-    expected_resolver_block_locals(
-        statements,
-        expr.as_deref(),
-        &mut scope_cursor,
-        &locals,
-        &mut expected,
-    );
-
-    assert!(expected.iter().any(|(name, _)| name == "value"));
-    assert!(expected.iter().any(|(name, _)| name == "input"));
-}
-
-#[test]
-fn expected_resolver_statement_locals_preserve_mutable_handoff() {
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let mut locals = scope_cursor.new_scope();
-    locals.insert("value".to_string(), true);
-    let mut expected = HashSet::new();
-
-    if resolver_var_decl_binds_local("value", false, false, &locals) {
-        expected_resolver_var_decl_local("value", false, &mut locals, &mut expected);
-    }
-
-    assert!(
-        expected.iter().all(|(name, _)| name != "value"),
-        "immutable declaration should reuse the mutable handoff binding"
-    );
-}
-
-#[test]
-fn expected_resolver_closure_locals_collects_params_and_body_bindings() {
-    let program = parse_program(
-        r#"
-main = () i32 {
-    mapper = (mut input: i32) i32 {
-        inner := input
-        inner
-    }
-    0
-}
-"#,
-    );
-    let Declaration::Function { body, .. } = &program.declarations[0] else {
-        panic!("expected function");
-    };
-    let Expression::Block { statements, .. } = body else {
-        panic!("expected block");
-    };
-    let Some(ast::Statement::VarDecl {
-        value:
-            Expression::Closure {
-                params,
-                body: closure_body,
-                ..
-            },
-        ..
-    }) = statements.first()
-    else {
-        panic!("expected closure var declaration");
-    };
-    let mut scope_cursor = ResolverScopeCursor::default();
-    let locals = scope_cursor.new_scope();
-    let mut expected = HashSet::new();
-
-    expected_resolver_closure_locals(
-        params,
-        closure_body,
-        &mut scope_cursor,
-        &locals,
-        &mut expected,
-    );
-
-    assert!(expected.iter().any(|(name, _)| name == "input"));
-    assert!(expected.iter().any(|(name, _)| name == "inner"));
-}
-
-#[test]
-fn resolver_declaration_metadata_tasks_collect_impl_blocks_with_declarations() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Option: Some(i32), None
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.impl = {
-    x_value = (value: Point) i32 { value.x }
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json)
-
-main = () i32 { 0 }
-"#,
-    );
-
-    let tasks = TypeChecker::collect_resolver_declaration_metadata_tasks(&program.declarations);
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-
-    assert_eq!(tasks.types.len(), 2);
-    assert_eq!(tasks.behaviors.len(), 1);
-    assert_eq!(tasks.behaviors[0].name, "Json");
-    assert_eq!(tasks.callable.len(), 2);
-    assert!(tasks.behavior_associations.extends.is_empty());
-    assert_eq!(tasks.behavior_associations.impls.len(), 1);
-    let behavior_impl = &tasks.behavior_associations.impls[0];
-    assert_eq!(behavior_impl.ast_type_name, "Point");
-    assert_eq!(behavior_impl.behavior, "Json");
-    assert_eq!(behavior_impl.methods.len(), 1);
-    assert_eq!(tasks.behavior_associations.requires.len(), 1);
-    let requires = &tasks.behavior_associations.requires[0];
-    assert_eq!(requires.type_name, "Point");
-    assert_eq!(requires.behavior, "Json");
-    assert_eq!(tasks.type_references.len(), 6);
-
-    let tc = TypeChecker::new();
-    let refresh_tasks = tc.resolver_type_behavior_refresh_tasks(&tasks, &symbols);
-    assert_eq!(refresh_tasks.len(), 2);
-    assert_eq!(refresh_tasks[0].restored_name, "Point");
-    assert_eq!(refresh_tasks[1].restored_name, "Option");
-}
-
-#[test]
-fn expected_behavior_edges_build_parent_edges_from_extends_together() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let expected = ExpectedBehaviorEdges::parents_from(&program);
-    let edge = &expected.edges_for("PrettyJson")[0];
-
-    assert_eq!(edge.display, "Json");
-    assert_eq!(edge.metadata.name, "Json");
-    assert_eq!(edge.metadata.type_args, Vec::<AstType>::new());
-}
-
-#[test]
-fn behavior_ref_role_validation_emits_selected_contains_diagnostics() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-Point.implements(PrettyJson) {
-    pretty = (value: Point) str { "pretty" }
-}
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let ty = symbols
-        .lookup(Namespace::Type, "Point")
-        .expect("type symbol");
-    let mut tc = TypeChecker::new();
-
-    tc.validate_resolver_behavior_ref_contains_for_role(
-        BehaviorRefRole::Impl,
-        ty,
-        "Point",
-        expected_behavior_edge("Json", &[AstType::Str]),
-        Span::dummy(),
-    );
-
-    assert!(tc.diagnostics.iter().any(|d| d.code == "E0236" && d.message.contains(
-            "resolver type symbol 'Point' has behavior impls 'PrettyJson', expected to include 'Json<str>'"
-        )));
-    assert!(tc.diagnostics.iter().any(|d| d.code == "E0247" && d.message.contains(
-            "resolver type symbol 'Point' has behavior impl refs 'PrettyJson', expected to include 'Json<str>'"
-        )));
-}
-
-#[test]
-fn behavior_association_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup(Namespace::Type, "Point")
-        .expect("type symbol");
-    let entries = BehaviorAssociationAbsenceValidation {
-        impl_name_code: "IMPL_NAMES",
-        impl_ref_code: "IMPL_REFS",
-        required_name_code: "REQUIRED_NAMES",
-        required_ref_code: "REQUIRED_REFS",
-    }
-    .entries(symbol);
-
-    assert_eq!(
-        entries,
-        [
-            AbsentMetadataEntry::new(true, "IMPL_NAMES", "behavior impls"),
-            AbsentMetadataEntry::new(true, "IMPL_REFS", "typed behavior impls"),
-            AbsentMetadataEntry::new(true, "REQUIRED_NAMES", "behavior requires"),
-            AbsentMetadataEntry::new(true, "REQUIRED_REFS", "typed behavior requires"),
-        ]
-    );
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_module_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0279");
-    assert_eq!(validation.impl_ref_code, "E0378");
-    assert_eq!(validation.required_name_code, "E0280");
-    assert_eq!(validation.required_ref_code, "E0379");
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_import_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0295");
-    assert_eq!(validation.impl_ref_code, "E0369");
-    assert_eq!(validation.required_name_code, "E0296");
-    assert_eq!(validation.required_ref_code, "E0370");
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_local_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0263");
-    assert_eq!(validation.impl_ref_code, "E0387");
-    assert_eq!(validation.required_name_code, "E0264");
-    assert_eq!(validation.required_ref_code, "E0388");
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_variant_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::variant_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0341");
-    assert_eq!(validation.impl_ref_code, "E0395");
-    assert_eq!(validation.required_name_code, "E0342");
-    assert_eq!(validation.required_ref_code, "E0396");
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_behavior_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::behavior_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0327");
-    assert_eq!(validation.impl_ref_code, "E0401");
-    assert_eq!(validation.required_name_code, "E0328");
-    assert_eq!(validation.required_ref_code, "E0402");
-}
-
-#[test]
-fn behavior_association_absence_validation_uses_value_resolver_codes() {
-    let validation = BehaviorAssociationAbsenceValidation::value_resolver_codes();
-
-    assert_eq!(validation.impl_name_code, "E0306");
-    assert_eq!(validation.impl_ref_code, "E0407");
-    assert_eq!(validation.required_name_code, "E0307");
-    assert_eq!(validation.required_ref_code, "E0408");
-}
-
-#[test]
-fn behavior_declaration_absence_validation_builds_entries() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-    let symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    let symbol = symbols
-        .lookup(Namespace::Behavior, "PrettyJson")
-        .expect("behavior symbol");
-    let entries = BehaviorDeclarationAbsenceValidation {
-        method_signature_code: "METHODS",
-        method_type_code: "TYPED_METHODS",
-        parent_name_code: "PARENTS",
-        parent_ref_code: "TYPED_PARENTS",
-    }
-    .entries(symbol);
-
-    assert_eq!(
-        entries,
-        [
-            AbsentMetadataEntry::new(true, "METHODS", "behavior methods"),
-            AbsentMetadataEntry::new(true, "TYPED_METHODS", "typed behavior methods"),
-            AbsentMetadataEntry::new(true, "PARENTS", "behavior parents"),
-            AbsentMetadataEntry::new(true, "TYPED_PARENTS", "typed behavior parents"),
-        ]
-    );
-}
-
-#[test]
-fn behavior_declaration_absence_validation_uses_module_resolver_codes() {
-    let validation = BehaviorDeclarationAbsenceValidation::module_resolver_codes();
-
-    assert_eq!(validation.method_signature_code, "E0277");
-    assert_eq!(validation.method_type_code, "E0376");
-    assert_eq!(validation.parent_name_code, "E0278");
-    assert_eq!(validation.parent_ref_code, "E0377");
-}
-
-#[test]
-fn behavior_declaration_absence_validation_uses_import_resolver_codes() {
-    let validation = BehaviorDeclarationAbsenceValidation::import_resolver_codes();
-
-    assert_eq!(validation.method_signature_code, "E0293");
-    assert_eq!(validation.method_type_code, "E0367");
-    assert_eq!(validation.parent_name_code, "E0294");
-    assert_eq!(validation.parent_ref_code, "E0368");
-}
-
-#[test]
-fn behavior_declaration_absence_validation_uses_local_resolver_codes() {
-    let validation = BehaviorDeclarationAbsenceValidation::local_resolver_codes();
-
-    assert_eq!(validation.method_signature_code, "E0261");
-    assert_eq!(validation.method_type_code, "E0385");
-    assert_eq!(validation.parent_name_code, "E0262");
-    assert_eq!(validation.parent_ref_code, "E0386");
-}
-
-#[test]
-fn behavior_declaration_absence_validation_uses_variant_resolver_codes() {
-    let validation = BehaviorDeclarationAbsenceValidation::variant_resolver_codes();
-
-    assert_eq!(validation.method_signature_code, "E0339");
-    assert_eq!(validation.method_type_code, "E0393");
-    assert_eq!(validation.parent_name_code, "E0340");
-    assert_eq!(validation.parent_ref_code, "E0394");
-}
-
-#[test]
-fn behavior_declaration_absence_validation_uses_value_resolver_codes() {
-    let validation = BehaviorDeclarationAbsenceValidation::value_resolver_codes();
-
-    assert_eq!(validation.method_signature_code, "E0304");
-    assert_eq!(validation.method_type_code, "E0405");
-    assert_eq!(validation.parent_name_code, "E0305");
-    assert_eq!(validation.parent_ref_code, "E0406");
-}
+mod resolver_validation;
 
 #[test]
 fn value_signature_absence_validation_builds_entries() {
@@ -14555,1270 +11286,4 @@ fn substitute_type_preserves_function_type_arguments_in_nested_generics() {
     );
 }
 
-#[test]
-fn generic_function_collection() {
-    use crate::ast::Expression;
-    let mut tc = TypeChecker::new();
-    let decls = vec![Declaration::Function {
-        name: "identity".into(),
-        type_params: vec![crate::ast::declarations::TypeParam {
-            name: "T".into(),
-            constraint: None,
-            constraint_type_args: Vec::new(),
-            span: Span::dummy(),
-        }],
-        params: vec![crate::ast::Param {
-            name: "x".into(),
-            ty: AstType::Named("T".into()),
-            mutable: false,
-            span: Span::dummy(),
-        }],
-        return_type: Some(AstType::Named("T".into())),
-        body: Expression::Block {
-            statements: Vec::new(),
-            expr: None,
-            span: Span::dummy(),
-        },
-        public: false,
-        span: Span::dummy(),
-    }];
-    tc.collect_declarations(&decls);
-    let info = tc.functions.get("identity").unwrap();
-    assert_eq!(info.type_params, vec!["T".to_string()]);
-}
-
-#[test]
-fn generic_method_collection() {
-    use crate::ast::Expression;
-    let mut tc = TypeChecker::new();
-    let decls = vec![Declaration::Method {
-        type_name: "Box".into(),
-        method_name: "get".into(),
-        type_params: vec![crate::ast::declarations::TypeParam {
-            name: "T".into(),
-            constraint: None,
-            constraint_type_args: Vec::new(),
-            span: Span::dummy(),
-        }],
-        params: vec![crate::ast::Param {
-            name: "value".into(),
-            ty: AstType::Named("T".into()),
-            mutable: false,
-            span: Span::dummy(),
-        }],
-        return_type: Some(AstType::Named("T".into())),
-        body: Expression::Block {
-            statements: Vec::new(),
-            expr: None,
-            span: Span::dummy(),
-        },
-        public: false,
-        span: Span::dummy(),
-    }];
-    tc.collect_declarations(&decls);
-    let info = tc.methods.get("Box.get").unwrap();
-    assert_eq!(info.type_params, vec!["T".to_string()]);
-    assert!(tc.generic_methods.contains_key("Box.get"));
-}
-
-#[test]
-fn type_impl_method_collection() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Point.impl = {
-    get = (self: Point) i32 {
-        self.x
-    }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.collect_declarations(&program.declarations);
-    let info = tc.methods.get("Point.get").unwrap();
-    assert_eq!(info.params.len(), 1);
-    assert_eq!(info.return_type, AstType::I32);
-}
-
-#[test]
-fn behavior_declaration_collection() {
-    let program = parse_program(
-        r#"
-Serializable: behavior {
-    to_json: (Self) String
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.collect_declarations(&program.declarations);
-    let info = tc.behaviors.get("Serializable").unwrap();
-    assert_eq!(info.name, "Serializable");
-    assert_eq!(info.methods.len(), 1);
-    assert_eq!(info.methods[0].name, "to_json");
-}
-
-#[test]
-fn behavior_impl_with_required_method_passes() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("valid behavior impl should typecheck");
-}
-
-#[test]
-fn behavior_impl_missing_required_method_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("missing behavior method should fail");
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "type `Point` implementation of `Json` is missing required method `to_json`"
-        )),
-        "expected missing behavior method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_can_omit_default_method() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str { "{}" }
-}
-
-Point.implements(Json) {
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("behavior impl may omit a method with a default body");
-}
-
-#[test]
-fn behavior_impl_duplicate_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("duplicate behavior impl should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("duplicate implementation of behavior `Json` for type `Point`")),
-        "expected duplicate behavior impl diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_generic_behavior_without_type_args_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) str
-}
-
-Point.implements(Json) {
-    encode = (value: Point) str { "point" }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior impl without type arguments should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic behavior `Json` expects 1 type arguments, found 0")),
-        "expected generic behavior impl arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_generic_behavior_with_type_args_passes_requires() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("generic behavior impl should satisfy matching generic requires");
-}
-
-#[test]
-fn behavior_impl_generic_behavior_type_arg_bound_failure_is_error() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Serializable<Point>) {
-    serialize = (value: Point) Point { value }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior type argument bound should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json<Point>` required by `T`")),
-        "expected generic behavior type argument bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_generic_behavior_type_arg_bound_passes_when_satisfied() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-
-Point: { x: i32 }
-
-Point.implements(Json<Point>) {
-    encode = (value: Point) Point { value }
-}
-
-Point.implements(Serializable<Point>) {
-    serialize = (value: Point) Point { value }
-}
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("generic behavior type argument bound should pass when satisfied");
-}
-
-#[test]
-fn behavior_requires_generic_behavior_type_arg_arity_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.requires(Json<i32, str>)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior requires arity mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic behavior `Json` expects 1 type arguments, found 2")),
-        "expected generic behavior requires arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_generic_behavior_substitutes_method_signature() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) i32 { 1 }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior impl return mismatch should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("method `encode` for behavior `Json_str` expects return `str`, found `i32`")),
-        "expected substituted behavior method return diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_overlapping_inherited_behavior_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-
-Point.implements(PrettyJson) {
-    to_json = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("overlapping inherited behavior impl should fail");
-    assert!(
-        errors.iter().any(|d| {
-            d.message.contains(
-                "overlapping implementations of behaviors `Json` and `PrettyJson` for type `Point`",
-            )
-        }),
-        "expected overlapping behavior impl diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_requires_passes_when_impl_exists() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-
-Point.requires(Json)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("requires should pass when behavior impl exists");
-}
-
-#[test]
-fn behavior_requires_rejects_missing_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.requires(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("requires should fail without behavior impl");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement required behavior `Json`")),
-        "expected requires missing impl diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_requires_generic_behavior_without_type_args_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) str
-}
-
-Point.requires(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior requires without type arguments should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic behavior `Json` expects 1 type arguments, found 0")),
-        "expected generic behavior requires arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_requires_parent_methods() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-
-Point.implements(PrettyJson) {
-    pretty = (value: Point) str { "pretty" }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("extended behavior should require parent methods");
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "type `Point` implementation of `PrettyJson` is missing required method `to_json`"
-        )),
-        "expected inherited missing method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_impl_satisfies_parent_requires() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-
-Point.implements(PrettyJson) {
-    to_json = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-
-Point.requires(Json)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("implementation of child behavior should satisfy parent requires");
-}
-
-#[test]
-fn behavior_extends_generic_parent_requires_substituted_methods() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(PrettyJson) {
-    pretty = (value: Point) str { "pretty" }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic parent method should be required with substituted signature");
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "type `Point` implementation of `PrettyJson` is missing required method `encode`"
-        )),
-        "expected inherited generic parent missing method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_generic_parent_satisfies_specialized_requires() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(PrettyJson) {
-    encode = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-
-Point.requires(Json<str>)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("child behavior impl should satisfy specialized generic parent requires");
-}
-
-#[test]
-fn behavior_extends_generic_parent_accepts_child_type_parameter_arg() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Serializable<T: Json<T>>: behavior {
-    serialize: (Self) T
-}
-
-Pretty<T: Json<T>>: behavior {
-    pretty: (Self) T
-}
-
-Pretty.extends(Serializable<T>)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("generic behavior parent should accept child type parameter args");
-}
-
-#[test]
-fn behavior_impl_generic_parent_overlap_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.implements(PrettyJson) {
-    encode = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("specialized parent and child behavior impls should overlap");
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "overlapping implementations of behaviors `Json_str` and `PrettyJson` for type `Point`"
-        )),
-        "expected specialized behavior impl overlap diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_distinct_generic_specializations_do_not_overlap() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-Point.implements(Json<i32>) {
-    encode = (value: Point) i32 { value.x }
-}
-
-Point.requires(Json<str>)
-Point.requires(Json<i32>)
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("distinct behavior specializations should not overlap");
-}
-
-#[test]
-fn behavior_extends_cycle_is_error() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-Json.extends(PrettyJson)
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("cyclic behavior inheritance should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("behavior inheritance cycle")),
-        "expected behavior inheritance cycle diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_duplicate_parent_is_error() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("duplicate behavior inheritance edge should fail");
-    assert!(
-        errors.iter().any(|d| {
-            d.message
-                .contains("duplicate behavior inheritance `PrettyJson.extends(Json)`")
-        }),
-        "expected duplicate behavior inheritance diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_duplicate_generic_parent_is_error() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json<str>)
-PrettyJson.extends(Json<str>)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("duplicate specialized behavior inheritance edge should fail");
-    assert!(
-        errors.iter().any(|d| {
-            d.message
-                .contains("duplicate behavior inheritance `PrettyJson.extends(Json<str>)`")
-        }),
-        "expected duplicate generic behavior inheritance diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_generic_parent_without_type_args_is_error() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    encode: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior extends parent without type arguments should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic behavior `Json` expects 1 type arguments, found 0")),
-        "expected generic behavior extends parent arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_extends_conflicting_method_signature_is_error() {
-    let program = parse_program(
-        r#"
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    to_json: (Self) i32
-}
-
-PrettyJson.extends(Json)
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("conflicting inherited behavior method should fail");
-    assert!(
-        errors.iter().any(|d| {
-            d.message
-                .contains("conflicting behavior method `to_json` inherited by `PrettyJson`")
-        }),
-        "expected conflicting inherited behavior method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_signature_mismatch_is_error() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: i32) i32 { value }
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("behavior impl signature mismatch should fail");
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("parameter 1 for method `to_json`")),
-        "expected behavior parameter mismatch diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("expects return `str`, found `i32`")),
-        "expected behavior return mismatch diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_explicit_type_arg_arity_is_error() {
-    let program = parse_program(
-        r#"
-identity<T> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    identity<i32, str>(1)
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("wrong generic type-argument arity should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic function `identity` expects 1 type arguments, found 2")),
-        "expected generic arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn nongeneric_function_explicit_type_args_are_error() {
-    let program = parse_program(
-        r#"
-id = (value: i32) i32 {
-    value
-}
-
-main = () i32 {
-    id<i32>(1)
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("non-generic function type arguments should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("non-generic function `id` does not accept type arguments")),
-        "expected non-generic type-argument diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_inference_failure_is_error() {
-    let program = parse_program(
-        r#"
-make_default<T> = () T {
-    0
-}
-
-main = () i32 {
-    make_default()
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("uninferred generic type argument should fail");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("cannot infer type argument `T` for generic function `make_default`")),
-        "expected generic inference diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_bound_references_unknown_behavior_is_error() {
-    let program = parse_program(
-        r#"
-show<T: Display> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    show(1)
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("unknown generic behavior bounds should fail");
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "generic bound `Display` on type parameter `T` references undefined behavior"
-        )),
-        "expected generic bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_bound_rejects_unspecialized_generic_behavior() {
-    let program = parse_program(
-        r#"
-Json<T>: behavior {
-    to_json: (Self) str
-}
-
-encode<T: Json> = (value: T) str {
-    "encoded"
-}
-
-main = () i32 {
-    0
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("generic behavior bound without type arguments should fail");
-    assert!(
-        errors.iter().any(|d| {
-            d.message
-                .contains("generic behavior `Json` expects 1 type arguments, found 0")
-        }),
-        "expected generic behavior bound arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_behavior_bound_with_type_args_accepts_matching_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<Point>) {
-    encode = (value: Point) Point { value }
-}
-
-identity<T: Json<T>> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    p = Point { x: 1 }
-    same = identity(p)
-    same.x
-}
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("generic behavior bound type argument should substitute at call site");
-}
-
-#[test]
-fn generic_behavior_bound_with_type_args_rejects_mismatched_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point.implements(Json<str>) {
-    encode = (value: Point) str { "point" }
-}
-
-identity<T: Json<T>> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    p = Point { x: 1 }
-    same = identity(p)
-    same.x
-}
-"#,
-    );
-
-    let errors = TypeChecker::new()
-        .check_program(&program)
-        .expect_err("generic behavior bound should require matching behavior type args");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json<Point>` required by `T`")),
-        "expected generic behavior bound type argument diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_generic_bound_accepts_later_behavior_declaration() {
-    let program = parse_program(
-        r#"
-Serializable<T: Json>: behavior {
-    encode: (Self) str
-}
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-main = () i32 {
-    0
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("behavior generic bounds should be independent of declaration order");
-}
-
-#[test]
-fn behavior_generic_bound_unknown_behavior_reports_once() {
-    let program = parse_program(
-        r#"
-Serializable<T: Missing>: behavior {
-    encode: (Self) str
-}
-
-main = () i32 {
-    0
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("unknown behavior generic bound should fail");
-    let count = errors
-        .iter()
-        .filter(|d| {
-            d.message.contains(
-                "generic bound `Missing` on type parameter `T` references undefined behavior",
-            )
-        })
-        .count();
-    assert_eq!(
-        count, 1,
-        "expected one behavior generic bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_behavior_bound_accepts_type_with_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str { "point" }
-}
-
-encode<T: Json> = (value: T) str {
-    "encoded"
-}
-
-main = () i32 {
-    p = Point { x: 1 }
-    encoded = encode(p)
-    0
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    tc.check_program(&program)
-        .expect("type with behavior impl should satisfy generic bound");
-}
-
-#[test]
-fn generic_behavior_bound_accepts_inherited_behavior_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-PrettyJson: behavior {
-    pretty: (Self) str
-}
-
-PrettyJson.extends(Json)
-
-Point.implements(PrettyJson) {
-    to_json = (value: Point) str { "point" }
-    pretty = (value: Point) str { "pretty" }
-}
-
-encode<T: Json> = (value: T) str {
-    "encoded"
-}
-
-main = () i32 {
-    p = Point { x: 1 }
-    encoded = encode(p)
-    0
-}
-"#,
-    );
-
-    TypeChecker::new()
-        .check_program(&program)
-        .expect("child behavior impl should satisfy inherited generic bound");
-}
-
-#[test]
-fn generic_behavior_bound_rejects_type_without_impl() {
-    let program = parse_program(
-        r#"
-Point: { x: i32 }
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-encode<T: Json> = (value: T) str {
-    "encoded"
-}
-
-main = () i32 {
-    p = Point { x: 1 }
-    encoded = encode(p)
-    0
-}
-"#,
-    );
-
-    let mut tc = TypeChecker::new();
-    let errors = tc
-        .check_program(&program)
-        .expect_err("type without behavior impl should not satisfy generic bound");
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json`")),
-        "expected missing generic bound impl diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn func_info_non_generic_has_empty_type_params() {
-    use crate::ast::Expression;
-    let mut tc = TypeChecker::new();
-    let decls = vec![Declaration::Function {
-        name: "add".into(),
-        type_params: Vec::new(),
-        params: vec![
-            crate::ast::Param {
-                name: "a".into(),
-                ty: AstType::I32,
-                mutable: false,
-                span: Span::dummy(),
-            },
-            crate::ast::Param {
-                name: "b".into(),
-                ty: AstType::I32,
-                mutable: false,
-                span: Span::dummy(),
-            },
-        ],
-        return_type: Some(AstType::I32),
-        body: Expression::Block {
-            statements: Vec::new(),
-            expr: None,
-            span: Span::dummy(),
-        },
-        public: false,
-        span: Span::dummy(),
-    }];
-    tc.collect_declarations(&decls);
-    let info = tc.functions.get("add").unwrap();
-    assert!(info.type_params.is_empty());
-}
+mod generic_behaviors;
